🗺️ Atlas: [extract native sub-modules]
🕸️ Tangle: `crates/duke-interpreter/src/native.rs` had grown to nearly 38k lines long, representing a severe "Blob" anti-pattern that makes code difficult to navigate and maintain.

📐 Blueprint: Physically extracted domain-specific sets of native method implementations (`jul` for java.util.logging, `math`, and `stream`) into dedicated files within the `crates/duke-interpreter/src/native/` directory.

🧱 Stability: Preserves identical compilation semantics and boundaries by utilizing `include!` macros in the parent `native.rs` file, significantly reducing the size of the monolith while avoiding thousands of cascading private visibility scope compilation errors.

🔬 Verification: Builds successfully (`cargo check`) and all test suites strictly pass (`cargo test`).
