+++ crates/duke-interpreter/src/native.rs
@@ -317,1574 +317,7 @@ fn string_value_from_ref(heap: &duke_gc::Heap, string_ref: u64) -> Result<String
         .ok_or(Error::NullPointerException)
 }
 
-const JUL_LEVEL_VALUE_FIELD: usize = 0;
-
-const JUL_LOGGER_NAME_FIELD: usize = 0;
-const JUL_LOGGER_LEVEL_FIELD: usize = 1;
-const JUL_LOGGER_USE_PARENT_HANDLERS_FIELD: usize = 2;
-const JUL_LOGGER_PARENT_FIELD: usize = 3;
-const JUL_LOGGER_HANDLER_COUNT_FIELD: usize = 4;
-const JUL_LOGGER_HANDLERS_START: usize = 5;
-
-const JUL_HANDLER_LEVEL_FIELD: usize = 0;
-const JUL_HANDLER_FORMATTER_FIELD: usize = 1;
-
-const JUL_MANAGER_ROOT_LOGGER_FIELD: usize = 0;
-const JUL_MANAGER_LOGGER_COUNT_FIELD: usize = 1;
-const JUL_MANAGER_LOGGERS_START: usize = 2;
-
-const JUL_LOG_RECORD_LEVEL_FIELD: usize = 0;
-const JUL_LOG_RECORD_MESSAGE_FIELD: usize = 1;
-const JUL_LOG_RECORD_LOGGER_NAME_FIELD: usize = 2;
-const JUL_LOG_RECORD_THROWN_FIELD: usize = 3;
-const JUL_LOG_RECORD_MILLIS_FIELD: usize = 4;
-const JUL_LOG_RECORD_PARAMETERS_FIELD: usize = 5;
-
-const JUL_ENUM_INDEX_FIELD: usize = 0;
-const JUL_ENUM_COUNT_FIELD: usize = 1;
-const JUL_ENUM_NAMES_START: usize = 2;
-
-const JUL_LEVELS: [(&str, i32); 9] = [
-    ("SEVERE", 1000),
-    ("WARNING", 900),
-    ("INFO", 800),
-    ("CONFIG", 700),
-    ("FINE", 500),
-    ("FINER", 400),
-    ("FINEST", 300),
-    ("ALL", i32::MIN),
-    ("OFF", i32::MAX),
-];
-
-fn jul_level_spec_by_name(name: &str) -> Option<i32> {
-    JUL_LEVELS
-        .iter()
-        .find_map(|(level_name, value)| (*level_name == name).then_some(*value))
-}
-
-fn jul_level_name_by_value(value: i32) -> Option<&'static str> {
-    JUL_LEVELS
-        .iter()
-        .find_map(|(level_name, level_value)| (*level_value == value).then_some(*level_name))
-}
-
-fn jul_illegal_argument(message: String) -> Error {
-    push_pending_java_exception_message("java/lang/IllegalArgumentException", message);
-    Error::JavaException {
-        class_name: "java/lang/IllegalArgumentException".to_string(),
-    }
-}
-
-fn jul_allocate_level(heap: &mut duke_gc::Heap, name: &str, value: i32) -> Result<u64> {
-    let level_ref = heap.allocate("java/util/logging/Level".to_string(), 1);
-    heap.write_field(level_ref, JUL_LEVEL_VALUE_FIELD, Slot::Int(value))?;
-    heap.get_mut(level_ref)?.string_value = Some(name.to_string());
-    Ok(level_ref)
-}
-
-fn jul_level_ref_by_name(heap: &duke_gc::Heap, name: &str) -> Option<u64> {
-    heap.find_string_backed_object("java/util/logging/Level", name)
-}
-
-fn jul_level_ref_by_value(heap: &duke_gc::Heap, value: i32) -> Option<u64> {
-    let canonical_name = jul_level_name_by_value(value)?;
-    jul_level_ref_by_name(heap, canonical_name)
-}
-
-fn jul_level_slot_by_name(heap: &mut duke_gc::Heap, name: &str) -> Result<Slot> {
-    if let Some(level_ref) = jul_level_ref_by_name(heap, name) {
-        return Ok(Slot::Reference(Some(level_ref)));
-    }
-    let Some(value) = jul_level_spec_by_name(name) else {
-        return Err(jul_illegal_argument(format!("Bad level: {name}")));
-    };
-    Ok(Slot::Reference(Some(jul_allocate_level(heap, name, value)?)))
-}
-
-fn jul_level_slot_by_value(heap: &mut duke_gc::Heap, value: i32) -> Result<Slot> {
-    if let Some(level_ref) = jul_level_ref_by_value(heap, value) {
-        return Ok(Slot::Reference(Some(level_ref)));
-    }
-    let name = value.to_string();
-    Ok(Slot::Reference(Some(jul_allocate_level(heap, &name, value)?)))
-}
-
-fn jul_level_value(heap: &duke_gc::Heap, level_ref: u64) -> Result<i32> {
-    match heap.get(level_ref)?.fields.get(JUL_LEVEL_VALUE_FIELD) {
-        Some(Slot::Int(value)) => Ok(*value),
-        _ => Err(Error::InvalidRef { address: level_ref }),
-    }
-}
-
-fn jul_level_value_from_slot(heap: &duke_gc::Heap, level_slot: Slot) -> Result<i32> {
-    match level_slot {
-        Slot::Reference(Some(level_ref)) => jul_level_value(heap, level_ref),
-        Slot::Reference(None) => Err(Error::NullPointerException),
-        _ => Err(Error::TypeMismatch {
-            expected: "Level",
-            got: "other",
-        }),
-    }
-}
-
-fn jul_level_name(heap: &duke_gc::Heap, level_ref: u64) -> Result<String> {
-    if let Some(name) = heap.get(level_ref)?.string_value.clone() {
-        return Ok(name);
-    }
-    Ok(jul_level_value(heap, level_ref)?.to_string())
-}
-
-fn jul_level_name_from_slot(heap: &duke_gc::Heap, level_slot: Slot) -> Result<String> {
-    match level_slot {
-        Slot::Reference(Some(level_ref)) => jul_level_name(heap, level_ref),
-        Slot::Reference(None) => Ok("null".to_string()),
-        _ => Ok(jul_level_value_from_slot(heap, level_slot)?.to_string()),
-    }
-}
-
-fn jul_push_dynamic_field(heap: &mut duke_gc::Heap, obj_ref: u64, value: Slot) -> Result<()> {
-    let field_idx = {
-        let obj = heap.get_mut(obj_ref)?;
-        let field_idx = obj.fields.len();
-        obj.fields.push(Slot::Reference(None));
-        field_idx
-    };
-    heap.write_field(obj_ref, field_idx, value)
-}
-
-fn jul_create_simple_formatter(heap: &mut duke_gc::Heap) -> Slot {
-    Slot::Reference(Some(heap.allocate(
-        "java/util/logging/SimpleFormatter".to_string(),
-        0,
-    )))
-}
-
-fn jul_create_console_handler(heap: &mut duke_gc::Heap) -> Result<Slot> {
-    let handler_ref = heap.allocate("java/util/logging/ConsoleHandler".to_string(), 2);
-    let all_level = jul_level_slot_by_name(heap, "ALL")?;
-    let formatter = jul_create_simple_formatter(heap);
-    heap.write_field(handler_ref, JUL_HANDLER_LEVEL_FIELD, all_level)?;
-    heap.write_field(handler_ref, JUL_HANDLER_FORMATTER_FIELD, formatter)?;
-    Ok(Slot::Reference(Some(handler_ref)))
-}
-
-fn jul_create_logger(
-    heap: &mut duke_gc::Heap,
-    name: Option<&str>,
-    level_slot: Slot,
-    use_parent_handlers: bool,
-    parent_slot: Slot,
-    handlers: &[Slot],
-) -> Result<Slot> {
-    let logger_ref = heap.allocate(
-        "java/util/logging/Logger".to_string(),
-        JUL_LOGGER_HANDLERS_START + handlers.len(),
-    );
-    let name_slot = name.map_or(Slot::Reference(None), |logger_name| {
-        Slot::Reference(Some(heap.allocate_string(logger_name.to_string())))
-    });
-    heap.write_field(logger_ref, JUL_LOGGER_NAME_FIELD, name_slot)?;
-    heap.write_field(logger_ref, JUL_LOGGER_LEVEL_FIELD, level_slot)?;
-    heap.write_field(
-        logger_ref,
-        JUL_LOGGER_USE_PARENT_HANDLERS_FIELD,
-        Slot::Int(i32::from(use_parent_handlers)),
-    )?;
-    heap.write_field(logger_ref, JUL_LOGGER_PARENT_FIELD, parent_slot)?;
-    heap.write_field(
-        logger_ref,
-        JUL_LOGGER_HANDLER_COUNT_FIELD,
-        Slot::Int(i32::try_from(handlers.len()).unwrap_or(i32::MAX)),
-    )?;
-    for (idx, handler) in handlers.iter().copied().enumerate() {
-        heap.write_field(logger_ref, JUL_LOGGER_HANDLERS_START + idx, handler)?;
-    }
-    Ok(Slot::Reference(Some(logger_ref)))
-}
-
-fn jul_bootstrap_fallback_manager(heap: &mut duke_gc::Heap) -> Result<u64> {
-    let info_level = jul_level_slot_by_name(heap, "INFO")?;
-    let root_handler = jul_create_console_handler(heap)?;
-    let root_logger = jul_create_logger(
-        heap,
-        Some(""),
-        info_level,
-        false,
-        Slot::Reference(None),
-        &[root_handler],
-    )?;
-    let global_logger = jul_create_logger(
-        heap,
-        Some("global"),
-        Slot::Reference(None),
-        true,
-        root_logger,
-        &[],
-    )?;
-    let manager_ref = heap.allocate("java/util/logging/LogManager".to_string(), 2);
-    heap.get_mut(manager_ref)?.string_value = Some("singleton".to_string());
-    heap.write_field(manager_ref, JUL_MANAGER_ROOT_LOGGER_FIELD, root_logger)?;
-    heap.write_field(manager_ref, JUL_MANAGER_LOGGER_COUNT_FIELD, Slot::Int(0))?;
-    jul_manager_insert_logger(heap, manager_ref, "", root_logger)?;
-    jul_manager_insert_logger(heap, manager_ref, "global", global_logger)?;
-    Ok(manager_ref)
-}
-
-fn jul_manager_ref(heap: &mut duke_gc::Heap) -> Result<u64> {
-    if let Some(manager_ref) =
-        heap.find_string_backed_object("java/util/logging/LogManager", "singleton")
-    {
-        return Ok(manager_ref);
-    }
-    jul_bootstrap_fallback_manager(heap)
-}
-
-fn jul_manager_count(heap: &duke_gc::Heap, manager_ref: u64) -> usize {
-    match heap
-        .get(manager_ref)
-        .ok()
-        .and_then(|manager| manager.fields.get(JUL_MANAGER_LOGGER_COUNT_FIELD).copied())
-    {
-        Some(Slot::Int(count)) => usize::try_from(count.max(0)).unwrap_or(0),
-        _ => 0,
-    }
-}
-
-fn jul_manager_find_logger_by_name(
-    heap: &duke_gc::Heap,
-    manager_ref: u64,
-    name: &str,
-) -> Result<Option<u64>> {
-    let fields = heap.get(manager_ref)?.fields.clone();
-    let count = jul_manager_count(heap, manager_ref);
-    for idx in 0..count {
-        let name_idx = JUL_MANAGER_LOGGERS_START + idx * 2;
-        let logger_idx = name_idx + 1;
-        let Some(Slot::Reference(Some(name_ref))) = fields.get(name_idx).copied() else {
-            continue;
-        };
-        if string_value_from_ref(heap, name_ref)? == name
-            && let Some(Slot::Reference(Some(logger_ref))) = fields.get(logger_idx).copied()
-        {
-            return Ok(Some(logger_ref));
-        }
-    }
-    Ok(None)
-}
-
-fn jul_manager_insert_logger(
-    heap: &mut duke_gc::Heap,
-    manager_ref: u64,
-    name: &str,
-    logger_slot: Slot,
-) -> Result<bool> {
-    if jul_manager_find_logger_by_name(heap, manager_ref, name)?.is_some() {
-        return Ok(false);
-    }
-    let count = jul_manager_count(heap, manager_ref);
-    let name_ref = heap.allocate_string(name.to_string());
-    jul_push_dynamic_field(heap, manager_ref, Slot::Reference(Some(name_ref)))?;
-    jul_push_dynamic_field(heap, manager_ref, logger_slot)?;
-    heap.write_field(
-        manager_ref,
-        JUL_MANAGER_LOGGER_COUNT_FIELD,
-        Slot::Int(i32::try_from(count.saturating_add(1)).unwrap_or(i32::MAX)),
-    )?;
-    Ok(true)
-}
-
-fn jul_root_logger_slot(heap: &mut duke_gc::Heap) -> Result<Slot> {
-    let manager_ref = jul_manager_ref(heap)?;
-    Ok(heap
-        .get(manager_ref)?
-        .fields
-        .get(JUL_MANAGER_ROOT_LOGGER_FIELD)
-        .copied()
-        .unwrap_or(Slot::Reference(None)))
-}
-
-fn jul_get_or_create_logger(heap: &mut duke_gc::Heap, name: &str) -> Result<u64> {
-    let manager_ref = jul_manager_ref(heap)?;
-    if let Some(logger_ref) = jul_manager_find_logger_by_name(heap, manager_ref, name)? {
-        return Ok(logger_ref);
-    }
-    let parent_slot = jul_root_logger_slot(heap)?;
-    let logger_slot = jul_create_logger(
-        heap,
-        Some(name),
-        Slot::Reference(None),
-        true,
-        parent_slot,
-        &[],
-    )?;
-    jul_manager_insert_logger(heap, manager_ref, name, logger_slot)?;
-    match logger_slot {
-        Slot::Reference(Some(logger_ref)) => Ok(logger_ref),
-        _ => Err(Error::InvalidRef { address: 0 }),
-    }
-}
-
-fn jul_logger_name_string(heap: &duke_gc::Heap, logger_ref: u64) -> Result<Option<String>> {
-    match heap
-        .get(logger_ref)?
-        .fields
-        .get(JUL_LOGGER_NAME_FIELD)
-        .copied()
-        .unwrap_or(Slot::Reference(None))
-    {
-        Slot::Reference(Some(name_ref)) => Ok(Some(string_value_from_ref(heap, name_ref)?)),
-        Slot::Reference(None) => Ok(None),
-        _ => Err(Error::TypeMismatch {
-            expected: "String",
-            got: "other",
-        }),
-    }
-}
-
-fn jul_effective_level_value(heap: &duke_gc::Heap, logger_ref: u64) -> Result<i32> {
-    let mut current = Some(logger_ref);
-    let mut depth = 0usize;
-    while let Some(current_ref) = current {
-        if depth > 64 {
-            return Ok(800);
-        }
-        let logger = heap.get(current_ref)?;
-        let level_slot = logger
-            .fields
-            .get(JUL_LOGGER_LEVEL_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None));
-        if !matches!(level_slot, Slot::Reference(None)) {
-            return jul_level_value_from_slot(heap, level_slot);
-        }
-        current = match logger
-            .fields
-            .get(JUL_LOGGER_PARENT_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None))
-        {
-            Slot::Reference(Some(parent_ref)) => Some(parent_ref),
-            _ => None,
-        };
-        depth = depth.saturating_add(1);
-    }
-    Ok(800)
-}
-
-const fn jul_is_loggable_for_value(effective: i32, candidate: i32) -> bool {
-    effective != i32::MAX && candidate != i32::MAX && candidate >= effective
-}
-
-fn jul_logger_is_loggable(heap: &duke_gc::Heap, logger_ref: u64, level_slot: Slot) -> Result<bool> {
-    let candidate = jul_level_value_from_slot(heap, level_slot)?;
-    let effective = jul_effective_level_value(heap, logger_ref)?;
-    Ok(jul_is_loggable_for_value(effective, candidate))
-}
-
-fn jul_slot_to_text(heap: &duke_gc::Heap, slot: Slot) -> Result<String> {
-    match slot {
-        Slot::Reference(Some(obj_ref)) => Ok(heap_object_to_string(heap.get(obj_ref)?, obj_ref)),
-        Slot::Reference(None) => Ok("null".to_string()),
-        Slot::Int(value) => Ok(value.to_string()),
-        Slot::Long(value) => Ok(value.to_string()),
-        Slot::Float(value) => Ok(format_java_float(value)),
-        Slot::Double(value) => Ok(format_java_double(value)),
-        Slot::ReturnAddress(value) => Ok(value.to_string()),
-    }
-}
-
-fn jul_apply_parameters(heap: &duke_gc::Heap, message: &str, params: &[Slot]) -> Result<String> {
-    let mut rendered = message.to_string();
-    for (idx, param) in params.iter().copied().enumerate() {
-        let token = format!("{{{idx}}}");
-        if rendered.contains(&token) {
-            let value = jul_slot_to_text(heap, param)?;
-            rendered = rendered.replace(&token, &value);
-        }
-    }
-    Ok(rendered)
-}
-
-fn jul_current_millis() -> i64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map_or(0, |duration| {
-            i64::try_from(duration.as_millis()).unwrap_or(i64::MAX)
-        })
-}
-
-fn jul_allocate_record(
-    heap: &mut duke_gc::Heap,
-    logger_ref: u64,
-    level_slot: Slot,
-    message: &str,
-    thrown_slot: Slot,
-    params: &[Slot],
-) -> Result<u64> {
-    let record_ref = heap.allocate("java/util/logging/LogRecord".to_string(), 6);
-    let message_ref = heap.allocate_string(message.to_string());
-    let logger_name_slot = heap
-        .get(logger_ref)?
-        .fields
-        .get(JUL_LOGGER_NAME_FIELD)
-        .copied()
-        .unwrap_or(Slot::Reference(None));
-    let parameters_slot = if params.is_empty() {
-        Slot::Reference(None)
-    } else {
-        let array_ref = allocate_reference_array_from_slots(heap, "[Ljava/lang/Object;", params)?;
-        Slot::Reference(Some(array_ref))
-    };
-    heap.write_field(record_ref, JUL_LOG_RECORD_LEVEL_FIELD, level_slot)?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_MESSAGE_FIELD,
-        Slot::Reference(Some(message_ref)),
-    )?;
-    heap.write_field(record_ref, JUL_LOG_RECORD_LOGGER_NAME_FIELD, logger_name_slot)?;
-    heap.write_field(record_ref, JUL_LOG_RECORD_THROWN_FIELD, thrown_slot)?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_MILLIS_FIELD,
-        Slot::Long(jul_current_millis()),
-    )?;
-    heap.write_field(record_ref, JUL_LOG_RECORD_PARAMETERS_FIELD, parameters_slot)?;
-    Ok(record_ref)
-}
-
-fn jul_record_level_slot(heap: &duke_gc::Heap, record_ref: u64) -> Slot {
-    heap.get(record_ref)
-        .ok()
-        .and_then(|record| record.fields.get(JUL_LOG_RECORD_LEVEL_FIELD).copied())
-        .unwrap_or(Slot::Reference(None))
-}
-
-fn jul_record_message(heap: &duke_gc::Heap, record_ref: u64) -> Result<String> {
-    match heap
-        .get(record_ref)?
-        .fields
-        .get(JUL_LOG_RECORD_MESSAGE_FIELD)
-        .copied()
-        .unwrap_or(Slot::Reference(None))
-    {
-        Slot::Reference(Some(message_ref)) => string_value_from_ref(heap, message_ref),
-        Slot::Reference(None) => Ok(String::new()),
-        _ => Err(Error::TypeMismatch {
-            expected: "String",
-            got: "other",
-        }),
-    }
-}
-
-fn jul_format_record(heap: &duke_gc::Heap, record_ref: u64) -> Result<String> {
-    let level_name = jul_level_name_from_slot(heap, jul_record_level_slot(heap, record_ref))?;
-    let message = jul_record_message(heap, record_ref)?;
-    Ok(format!("{level_name}: {message}"))
-}
-
-fn jul_handler_allows_record(
-    heap: &duke_gc::Heap,
-    handler_ref: u64,
-    record_ref: u64,
-) -> Result<bool> {
-    let handler_level = heap
-        .get(handler_ref)?
-        .fields
-        .get(JUL_HANDLER_LEVEL_FIELD)
-        .copied()
-        .unwrap_or(Slot::Reference(None));
-    if matches!(handler_level, Slot::Reference(None)) {
-        return Ok(true);
-    }
-    let handler_value = jul_level_value_from_slot(heap, handler_level)?;
-    let record_value = jul_level_value_from_slot(heap, jul_record_level_slot(heap, record_ref))?;
-    Ok(jul_is_loggable_for_value(handler_value, record_value))
-}
-
-fn jul_publish_record_to_handler(
-    heap: &duke_gc::Heap,
-    out: &mut dyn Write,
-    handler_ref: u64,
-    record_ref: u64,
-) -> Result<()> {
-    if !jul_handler_allows_record(heap, handler_ref, record_ref)? {
-        return Ok(());
-    }
-    let line = jul_format_record(heap, record_ref)?;
-    writeln!(out, "{line}").ok();
-    Ok(())
-}
-
-fn jul_publish_record_to_logger(
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    logger_ref: u64,
-    record_ref: u64,
-    depth: usize,
-) -> Result<()> {
-    if depth > 64 {
-        return Ok(());
-    }
-    let fields = heap.get(logger_ref)?.fields.clone();
-    let handler_count = match fields.get(JUL_LOGGER_HANDLER_COUNT_FIELD) {
-        Some(Slot::Int(count)) => usize::try_from((*count).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    for idx in 0..handler_count {
-        if let Some(Slot::Reference(Some(handler_ref))) =
-            fields.get(JUL_LOGGER_HANDLERS_START + idx).copied()
-        {
-            jul_publish_record_to_handler(heap, out, handler_ref, record_ref)?;
-        }
-    }
-    let use_parent_handlers = matches!(
-        fields.get(JUL_LOGGER_USE_PARENT_HANDLERS_FIELD),
-        Some(Slot::Int(value)) if *value != 0
-    );
-    if use_parent_handlers
-        && let Some(Slot::Reference(Some(parent_ref))) = fields.get(JUL_LOGGER_PARENT_FIELD).copied()
-    {
-        jul_publish_record_to_logger(heap, out, parent_ref, record_ref, depth + 1)?;
-    }
-    Ok(())
-}
-
-fn jul_log_text(
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    logger_ref: u64,
-    level_slot: Slot,
-    message: &str,
-    params: &[Slot],
-    thrown_slot: Slot,
-) -> Result<()> {
-    let rendered = jul_apply_parameters(heap, message, params)?;
-    let record_ref = jul_allocate_record(
-        heap,
-        logger_ref,
-        level_slot,
-        &rendered,
-        thrown_slot,
-        params,
-    )?;
-    jul_publish_record_to_logger(heap, out, logger_ref, record_ref, 0)
-}
-
-fn jul_log_message_slot(
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    logger_ref: u64,
-    level_slot: Slot,
-    message_slot: Slot,
-    params: &[Slot],
-    thrown_slot: Slot,
-) -> Result<()> {
-    if !jul_logger_is_loggable(heap, logger_ref, level_slot)? {
-        return Ok(());
-    }
-    let message = jul_slot_to_text(heap, message_slot)?;
-    jul_log_text(
-        heap,
-        out,
-        logger_ref,
-        level_slot,
-        &message,
-        params,
-        thrown_slot,
-    )
-}
-
-fn jul_log_named_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    level_name: &str,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let level_slot = jul_level_slot_by_name(heap, level_name)?;
-    let message_slot = extract_slot_arg(args, 1);
-    jul_log_message_slot(
-        heap,
-        out,
-        logger_ref,
-        level_slot,
-        message_slot,
-        &[],
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_level_int_value(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let level_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(Slot::Int(jul_level_value(heap, level_ref)?)))
-}
-
-pub(crate) fn native_jul_level_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let level_ref = extract_ref_arg(args, 0)?;
-    let name_ref = heap.allocate_string(jul_level_name(heap, level_ref)?);
-    Ok(Some(Slot::Reference(Some(name_ref))))
-}
-
-pub(crate) fn native_jul_level_parse(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let name_ref = extract_ref_arg(args, 0)?;
-    let text = string_value_from_ref(heap, name_ref)?;
-    if jul_level_spec_by_name(text.as_str()).is_some() {
-        return Ok(Some(jul_level_slot_by_name(heap, &text)?));
-    }
-    if let Ok(value) = text.parse::<i32>() {
-        return Ok(Some(jul_level_slot_by_value(heap, value)?));
-    }
-    Err(jul_illegal_argument(format!("Bad level: {text}")))
-}
-
-pub(crate) fn native_jul_logger_get_logger(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let name_ref = extract_ref_arg(args, 0)?;
-    let name = string_value_from_ref(heap, name_ref)?;
-    let logger_ref = jul_get_or_create_logger(heap, &name)?;
-    Ok(Some(Slot::Reference(Some(logger_ref))))
-}
-
-pub(crate) fn native_jul_logger_get_logger_with_bundle(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    native_jul_logger_get_logger(args, heap, out, control)
-}
-
-pub(crate) fn native_jul_logger_get_global(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = jul_get_or_create_logger(heap, "global")?;
-    Ok(Some(Slot::Reference(Some(logger_ref))))
-}
-
-pub(crate) fn native_jul_logger_get_anonymous_logger(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let parent_slot = jul_root_logger_slot(heap)?;
-    Ok(Some(jul_create_logger(
-        heap,
-        None,
-        Slot::Reference(None),
-        true,
-        parent_slot,
-        &[],
-    )?))
-}
-
-pub(crate) fn native_jul_logger_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(
-        heap.get(logger_ref)?
-            .fields
-            .get(JUL_LOGGER_NAME_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None)),
-    ))
-}
-
-pub(crate) fn native_jul_logger_get_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(
-        heap.get(logger_ref)?
-            .fields
-            .get(JUL_LOGGER_LEVEL_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None)),
-    ))
-}
-
-pub(crate) fn native_jul_logger_set_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    heap.write_field(logger_ref, JUL_LOGGER_LEVEL_FIELD, extract_slot_arg(args, 1))?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_is_loggable(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let level_slot = extract_slot_arg(args, 1);
-    Ok(Some(Slot::Int(i32::from(jul_logger_is_loggable(
-        heap, logger_ref, level_slot,
-    )?))))
-}
-
-pub(crate) fn native_jul_logger_log(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    jul_log_message_slot(
-        heap,
-        out,
-        logger_ref,
-        extract_slot_arg(args, 1),
-        extract_slot_arg(args, 2),
-        &[],
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_log_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let param = extract_slot_arg(args, 3);
-    jul_log_message_slot(
-        heap,
-        out,
-        logger_ref,
-        extract_slot_arg(args, 1),
-        extract_slot_arg(args, 2),
-        &[param],
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_log_object_array(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let params = match extract_slot_arg(args, 3) {
-        Slot::Reference(Some(array_ref)) => heap.get(array_ref)?.fields.clone(),
-        Slot::Reference(None) => Vec::new(),
-        _ => {
-            return Err(Error::TypeMismatch {
-                expected: "Object[]",
-                got: "other",
-            });
-        }
-    };
-    jul_log_message_slot(
-        heap,
-        out,
-        logger_ref,
-        extract_slot_arg(args, 1),
-        extract_slot_arg(args, 2),
-        &params,
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_log_throwable(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    jul_log_message_slot(
-        heap,
-        out,
-        logger_ref,
-        extract_slot_arg(args, 1),
-        extract_slot_arg(args, 2),
-        &[],
-        extract_slot_arg(args, 3),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_severe(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "SEVERE")
-}
-
-pub(crate) fn native_jul_logger_warning(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "WARNING")
-}
-
-pub(crate) fn native_jul_logger_info(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "INFO")
-}
-
-pub(crate) fn native_jul_logger_config(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "CONFIG")
-}
-
-pub(crate) fn native_jul_logger_fine(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "FINE")
-}
-
-pub(crate) fn native_jul_logger_finer(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "FINER")
-}
-
-pub(crate) fn native_jul_logger_finest(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_named_level(args, heap, out, "FINEST")
-}
-
-pub(crate) fn native_jul_logger_entering(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let level_slot = jul_level_slot_by_name(heap, "FINER")?;
-    if !jul_logger_is_loggable(heap, logger_ref, level_slot)? {
-        return Ok(None);
-    }
-    let source_class = jul_slot_to_text(heap, extract_slot_arg(args, 1))?;
-    let source_method = jul_slot_to_text(heap, extract_slot_arg(args, 2))?;
-    jul_log_text(
-        heap,
-        out,
-        logger_ref,
-        level_slot,
-        &format!("ENTRY {source_class}.{source_method}"),
-        &[],
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_exiting(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let level_slot = jul_level_slot_by_name(heap, "FINER")?;
-    if !jul_logger_is_loggable(heap, logger_ref, level_slot)? {
-        return Ok(None);
-    }
-    let source_class = jul_slot_to_text(heap, extract_slot_arg(args, 1))?;
-    let source_method = jul_slot_to_text(heap, extract_slot_arg(args, 2))?;
-    jul_log_text(
-        heap,
-        out,
-        logger_ref,
-        level_slot,
-        &format!("RETURN {source_class}.{source_method}"),
-        &[],
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_throwing(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let level_slot = jul_level_slot_by_name(heap, "FINER")?;
-    if !jul_logger_is_loggable(heap, logger_ref, level_slot)? {
-        return Ok(None);
-    }
-    let source_class = jul_slot_to_text(heap, extract_slot_arg(args, 1))?;
-    let source_method = jul_slot_to_text(heap, extract_slot_arg(args, 2))?;
-    jul_log_text(
-        heap,
-        out,
-        logger_ref,
-        level_slot,
-        &format!("THROW {source_class}.{source_method}"),
-        &[],
-        extract_slot_arg(args, 3),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_add_handler(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let handler_slot = match extract_slot_arg(args, 1) {
-        Slot::Reference(Some(_)) => extract_slot_arg(args, 1),
-        Slot::Reference(None) => return Err(Error::NullPointerException),
-        _ => {
-            return Err(Error::TypeMismatch {
-                expected: "Handler",
-                got: "other",
-            });
-        }
-    };
-    let count = match heap
-        .get(logger_ref)?
-        .fields
-        .get(JUL_LOGGER_HANDLER_COUNT_FIELD)
-    {
-        Some(Slot::Int(count)) => usize::try_from((*count).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    jul_push_dynamic_field(heap, logger_ref, handler_slot)?;
-    heap.write_field(
-        logger_ref,
-        JUL_LOGGER_HANDLER_COUNT_FIELD,
-        Slot::Int(i32::try_from(count.saturating_add(1)).unwrap_or(i32::MAX)),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_remove_handler(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let target = extract_slot_arg(args, 1);
-    let fields = heap.get(logger_ref)?.fields.clone();
-    let count = match fields.get(JUL_LOGGER_HANDLER_COUNT_FIELD) {
-        Some(Slot::Int(count)) => usize::try_from((*count).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    let mut retained = Vec::new();
-    let mut removed = false;
-    for idx in 0..count {
-        let handler_slot = fields
-            .get(JUL_LOGGER_HANDLERS_START + idx)
-            .copied()
-            .unwrap_or(Slot::Reference(None));
-        if !removed && handler_slot == target {
-            removed = true;
-        } else {
-            retained.push(handler_slot);
-        }
-    }
-    heap.get_mut(logger_ref)?
-        .fields
-        .truncate(JUL_LOGGER_HANDLERS_START);
-    for handler_slot in &retained {
-        jul_push_dynamic_field(heap, logger_ref, *handler_slot)?;
-    }
-    heap.write_field(
-        logger_ref,
-        JUL_LOGGER_HANDLER_COUNT_FIELD,
-        Slot::Int(i32::try_from(retained.len()).unwrap_or(i32::MAX)),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_get_handlers(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let fields = heap.get(logger_ref)?.fields.clone();
-    let count = match fields.get(JUL_LOGGER_HANDLER_COUNT_FIELD) {
-        Some(Slot::Int(count)) => usize::try_from((*count).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    let handlers: Vec<Slot> = (0..count)
-        .map(|idx| {
-            fields
-                .get(JUL_LOGGER_HANDLERS_START + idx)
-                .copied()
-                .unwrap_or(Slot::Reference(None))
-        })
-        .collect();
-    let array_ref = allocate_reference_array_from_slots(
-        heap,
-        "[Ljava/util/logging/Handler;",
-        &handlers,
-    )?;
-    Ok(Some(Slot::Reference(Some(array_ref))))
-}
-
-pub(crate) fn native_jul_logger_set_use_parent_handlers(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    let enabled = extract_int_arg(args, 1)?;
-    heap.write_field(
-        logger_ref,
-        JUL_LOGGER_USE_PARENT_HANDLERS_FIELD,
-        Slot::Int(i32::from(enabled != 0)),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_logger_get_use_parent_handlers(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(
-        heap.get(logger_ref)?
-            .fields
-            .get(JUL_LOGGER_USE_PARENT_HANDLERS_FIELD)
-            .copied()
-            .unwrap_or(Slot::Int(0)),
-    ))
-}
-
-pub(crate) fn native_jul_logger_get_parent(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let logger_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(
-        heap.get(logger_ref)?
-            .fields
-            .get(JUL_LOGGER_PARENT_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None)),
-    ))
-}
-
-pub(crate) fn native_jul_log_manager_get_log_manager(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(Slot::Reference(Some(jul_manager_ref(heap)?))))
-}
-
-pub(crate) fn native_jul_log_manager_get_logger(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let manager_ref = extract_ref_arg(args, 0)?;
-    let name_ref = extract_ref_arg(args, 1)?;
-    let name = string_value_from_ref(heap, name_ref)?;
-    Ok(Some(
-        jul_manager_find_logger_by_name(heap, manager_ref, &name)?
-            .map_or(Slot::Reference(None), |logger_ref| {
-                Slot::Reference(Some(logger_ref))
-            }),
-    ))
-}
-
-pub(crate) fn native_jul_log_manager_get_logger_names(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let manager_ref = extract_ref_arg(args, 0)?;
-    let fields = heap.get(manager_ref)?.fields.clone();
-    let count = jul_manager_count(heap, manager_ref);
-    let names: Vec<Slot> = (0..count)
-        .map(|idx| {
-            fields
-                .get(JUL_MANAGER_LOGGERS_START + idx * 2)
-                .copied()
-                .unwrap_or(Slot::Reference(None))
-        })
-        .collect();
-    let enumeration_ref = heap.allocate(
-        "duke/util/JulLoggerNameEnumeration".to_string(),
-        JUL_ENUM_NAMES_START + names.len(),
-    );
-    heap.write_field(enumeration_ref, JUL_ENUM_INDEX_FIELD, Slot::Int(0))?;
-    heap.write_field(
-        enumeration_ref,
-        JUL_ENUM_COUNT_FIELD,
-        Slot::Int(i32::try_from(names.len()).unwrap_or(i32::MAX)),
-    )?;
-    for (idx, name_slot) in names.iter().copied().enumerate() {
-        heap.write_field(enumeration_ref, JUL_ENUM_NAMES_START + idx, name_slot)?;
-    }
-    Ok(Some(Slot::Reference(Some(enumeration_ref))))
-}
-
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_jul_noop_void(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(None)
-}
-
-pub(crate) fn native_jul_log_manager_read_configuration_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 1)?;
-    let stream_class = heap.get(stream_ref)?.class_name.clone();
-    for _ in 0..1_048_576 {
-        match ops.invoke(
-            heap,
-            out,
-            &stream_class,
-            "read",
-            "()I",
-            vec![Slot::Reference(Some(stream_ref))],
-        ) {
-            Ok(Some(Slot::Int(value))) if value < 0 => break,
-            Ok(Some(Slot::Int(_))) => {}
-            Ok(_) | Err(Error::MethodNotFound { .. }) => break,
-            Err(err) => return Err(err),
-        }
-    }
-    Ok(None)
-}
-
-pub(crate) fn native_jul_log_manager_add_logger(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let manager_ref = extract_ref_arg(args, 0)?;
-    let logger_ref = extract_ref_arg(args, 1)?;
-    let Some(name) = jul_logger_name_string(heap, logger_ref)? else {
-        return Ok(Some(Slot::Int(0)));
-    };
-    let inserted = jul_manager_insert_logger(
-        heap,
-        manager_ref,
-        &name,
-        Slot::Reference(Some(logger_ref)),
-    )?;
-    Ok(Some(Slot::Int(i32::from(inserted))))
-}
-
-pub(crate) fn native_jul_handler_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    let all_level = jul_level_slot_by_name(heap, "ALL")?;
-    heap.write_field(handler_ref, JUL_HANDLER_LEVEL_FIELD, all_level)?;
-    heap.write_field(
-        handler_ref,
-        JUL_HANDLER_FORMATTER_FIELD,
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_console_handler_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    let all_level = jul_level_slot_by_name(heap, "ALL")?;
-    let formatter = jul_create_simple_formatter(heap);
-    heap.write_field(handler_ref, JUL_HANDLER_LEVEL_FIELD, all_level)?;
-    heap.write_field(handler_ref, JUL_HANDLER_FORMATTER_FIELD, formatter)?;
-    Ok(None)
-}
-
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_jul_handler_publish(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(None)
-}
-
-pub(crate) fn native_jul_console_handler_publish(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    let record_ref = extract_ref_arg(args, 1)?;
-    jul_publish_record_to_handler(heap, out, handler_ref, record_ref)?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_handler_set_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    heap.write_field(handler_ref, JUL_HANDLER_LEVEL_FIELD, extract_slot_arg(args, 1))?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_handler_get_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(
-        heap.get(handler_ref)?
-            .fields
-            .get(JUL_HANDLER_LEVEL_FIELD)
-            .copied()
-            .unwrap_or(Slot::Reference(None)),
-    ))
-}
-
-pub(crate) fn native_jul_handler_set_formatter(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let handler_ref = extract_ref_arg(args, 0)?;
-    heap.write_field(
-        handler_ref,
-        JUL_HANDLER_FORMATTER_FIELD,
-        extract_slot_arg(args, 1),
-    )?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_simple_formatter_format(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let record_ref = extract_ref_arg(args, 1)?;
-    let formatted_ref = heap.allocate_string(jul_format_record(heap, record_ref)?);
-    Ok(Some(Slot::Reference(Some(formatted_ref))))
-}
-
-pub(crate) fn native_jul_log_record_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let record_ref = extract_ref_arg(args, 0)?;
-    heap.write_field(record_ref, JUL_LOG_RECORD_LEVEL_FIELD, extract_slot_arg(args, 1))?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_MESSAGE_FIELD,
-        extract_slot_arg(args, 2),
-    )?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_LOGGER_NAME_FIELD,
-        Slot::Reference(None),
-    )?;
-    heap.write_field(record_ref, JUL_LOG_RECORD_THROWN_FIELD, Slot::Reference(None))?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_MILLIS_FIELD,
-        Slot::Long(jul_current_millis()),
-    )?;
-    heap.write_field(
-        record_ref,
-        JUL_LOG_RECORD_PARAMETERS_FIELD,
-        Slot::Reference(None),
-    )?;
-    Ok(None)
-}
-
-fn jul_log_record_get_field(
-    args: &[Slot],
-    heap: &duke_gc::Heap,
-    field_idx: usize,
-) -> Result<Slot> {
-    let record_ref = extract_ref_arg(args, 0)?;
-    Ok(heap
-        .get(record_ref)?
-        .fields
-        .get(field_idx)
-        .copied()
-        .unwrap_or(Slot::Reference(None)))
-}
-
-fn jul_log_record_set_field(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    field_idx: usize,
-    value_idx: usize,
-) -> Result<Option<Slot>> {
-    let record_ref = extract_ref_arg(args, 0)?;
-    heap.write_field(record_ref, field_idx, extract_slot_arg(args, value_idx))?;
-    Ok(None)
-}
-
-pub(crate) fn native_jul_log_record_get_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(jul_log_record_get_field(
-        args,
-        heap,
-        JUL_LOG_RECORD_LEVEL_FIELD,
-    )?))
-}
-
-pub(crate) fn native_jul_log_record_set_level(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_record_set_field(args, heap, JUL_LOG_RECORD_LEVEL_FIELD, 1)
-}
-
-pub(crate) fn native_jul_log_record_get_message(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(jul_log_record_get_field(
-        args,
-        heap,
-        JUL_LOG_RECORD_MESSAGE_FIELD,
-    )?))
-}
-
-pub(crate) fn native_jul_log_record_set_message(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_record_set_field(args, heap, JUL_LOG_RECORD_MESSAGE_FIELD, 1)
-}
-
-pub(crate) fn native_jul_log_record_get_logger_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(jul_log_record_get_field(
-        args,
-        heap,
-        JUL_LOG_RECORD_LOGGER_NAME_FIELD,
-    )?))
-}
-
-pub(crate) fn native_jul_log_record_set_logger_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_record_set_field(args, heap, JUL_LOG_RECORD_LOGGER_NAME_FIELD, 1)
-}
-
-pub(crate) fn native_jul_log_record_get_thrown(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(jul_log_record_get_field(
-        args,
-        heap,
-        JUL_LOG_RECORD_THROWN_FIELD,
-    )?))
-}
-
-pub(crate) fn native_jul_log_record_set_thrown(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_record_set_field(args, heap, JUL_LOG_RECORD_THROWN_FIELD, 1)
-}
-
-pub(crate) fn native_jul_log_record_get_millis(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let record_ref = extract_ref_arg(args, 0)?;
-    Ok(Some(match heap.get(record_ref)?.fields.get(JUL_LOG_RECORD_MILLIS_FIELD) {
-        Some(Slot::Long(value)) => Slot::Long(*value),
-        _ => Slot::Long(0),
-    }))
-}
-
-pub(crate) fn native_jul_log_record_get_parameters(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    Ok(Some(jul_log_record_get_field(
-        args,
-        heap,
-        JUL_LOG_RECORD_PARAMETERS_FIELD,
-    )?))
-}
-
-pub(crate) fn native_jul_log_record_set_parameters(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    jul_log_record_set_field(args, heap, JUL_LOG_RECORD_PARAMETERS_FIELD, 1)
-}
-
-pub(crate) fn native_jul_logger_names_has_more_elements(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let enum_ref = extract_ref_arg(args, 0)?;
-    let fields = &heap.get(enum_ref)?.fields;
-    let index = match fields.get(JUL_ENUM_INDEX_FIELD) {
-        Some(Slot::Int(index)) => *index,
-        _ => 0,
-    };
-    let count = match fields.get(JUL_ENUM_COUNT_FIELD) {
-        Some(Slot::Int(count)) => *count,
-        _ => 0,
-    };
-    Ok(Some(Slot::Int(i32::from(index < count))))
-}
-
-pub(crate) fn native_jul_logger_names_next_element(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let enum_ref = extract_ref_arg(args, 0)?;
-    let fields = heap.get(enum_ref)?.fields.clone();
-    let index = match fields.get(JUL_ENUM_INDEX_FIELD) {
-        Some(Slot::Int(index)) => usize::try_from((*index).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    let count = match fields.get(JUL_ENUM_COUNT_FIELD) {
-        Some(Slot::Int(count)) => usize::try_from((*count).max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    if index >= count {
-        return Ok(Some(Slot::Reference(None)));
-    }
-    heap.write_field(
-        enum_ref,
-        JUL_ENUM_INDEX_FIELD,
-        Slot::Int(i32::try_from(index.saturating_add(1)).unwrap_or(i32::MAX)),
-    )?;
-    Ok(Some(
-        fields
-            .get(JUL_ENUM_NAMES_START + index)
-            .copied()
-            .unwrap_or(Slot::Reference(None)),
-    ))
-}
-
+include!("native/jul.rs");
 const REPLACEMENT_CHAR: char = '\u{fffd}';
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -5430,2489 +3863,7 @@ pub(crate) fn native_collections_fill(
     Ok(None)
 }
 
-// ---- Stream natives ----
-// duke/util/Stream: fields[0]=Int(size), fields[1..]=element refs
-
-/// Native: `Stream.of(Object[])Stream` — create stream from varargs array.
-pub(crate) fn native_stream_of(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let arr_ref = extract_ref_arg(args, 0)?;
-    let elems: Vec<Slot> = heap.get(arr_ref)?.fields.clone();
-    let n = i32::try_from(elems.len()).unwrap_or(0);
-    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(n);
-    for elem in elems {
-        heap.get_mut(stream_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `ArrayList.stream()` — wrap `ArrayList` elements into a `Stream`.
-pub(crate) fn native_arraylist_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let list_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(list_ref)?.fields.first() {
-        Some(Slot::Int(n)) => *n,
-        _ => 0,
-    };
-    let elems: Vec<Slot> =
-        heap.get(list_ref)?.fields[1..=usize::try_from(size).unwrap_or(0)].to_vec();
-    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(size);
-    for elem in elems {
-        heap.get_mut(stream_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `Stream.count()J` — returns the number of elements as a long.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_stream_count(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let n = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => i64::from(*n),
-        _ => 0,
-    };
-    Ok(Some(Slot::Long(n)))
-}
-
-/// Native: `Stream.filter(Predicate)Stream` — keeps elements where `predicate.test()` returns true.
-pub(crate) fn native_stream_filter(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let pred_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(pred_ref)) = pred_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let pred_class = heap.get(pred_ref)?.class_name.clone();
-    let mut kept: Vec<Slot> = Vec::new();
-    for elem in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &pred_class,
-            "test",
-            "(Ljava/lang/Object;)Z",
-            vec![Slot::Reference(Some(pred_ref)), elem],
-        )?;
-        if matches!(result, Some(Slot::Int(1))) {
-            kept.push(elem);
-        }
-    }
-    let new_size = i32::try_from(kept.len()).unwrap_or(0);
-    let new_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(new_stream)?.fields[0] = Slot::Int(new_size);
-    for elem in kept {
-        heap.get_mut(new_stream)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(new_stream))))
-}
-
-/// Native: `Stream.map(Function)Stream` — transforms each element via `function.apply()`.
-pub(crate) fn native_stream_map(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let fn_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let mut mapped: Vec<Slot> = Vec::with_capacity(elems.len());
-    for elem in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &fn_class,
-            "apply",
-            "(Ljava/lang/Object;)Ljava/lang/Object;",
-            vec![Slot::Reference(Some(fn_ref)), elem],
-        )?;
-        // Box primitive results so stream elements are always References (Java type-erasure)
-        let boxed = box_primitive_slot(result.unwrap_or(Slot::Reference(None)), heap);
-        mapped.push(boxed);
-    }
-    let new_size = i32::try_from(mapped.len()).unwrap_or(0);
-    let new_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(new_stream)?.fields[0] = Slot::Int(new_size);
-    for elem in mapped {
-        heap.get_mut(new_stream)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(new_stream))))
-}
-
-/// Native: `Stream.forEach(Consumer)V` — calls `consumer.accept()` on each element.
-pub(crate) fn native_stream_for_each(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let consumer_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(consumer_ref)) = consumer_slot else {
-        return Ok(None);
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let consumer_class = heap.get(consumer_ref)?.class_name.clone();
-    for elem in elems {
-        ops.invoke(
-            heap,
-            out,
-            &consumer_class,
-            "accept",
-            "(Ljava/lang/Object;)V",
-            vec![Slot::Reference(Some(consumer_ref)), elem],
-        )?;
-    }
-    Ok(None)
-}
-
-/// Native: `Stream.collect(Collector)Object` — collects to list (only toList collector supported).
-#[allow(
-    clippy::too_many_lines,
-    clippy::only_used_in_recursion,
-    clippy::cognitive_complexity
-)]
-pub(crate) fn native_stream_collect(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => *n,
-        _ => 0,
-    };
-    let elems: Vec<Slot> =
-        heap.get(stream_ref)?.fields[1..=usize::try_from(size).unwrap_or(0)].to_vec();
-
-    // Dispatch on collector type.
-    let collector_class = match extract_ref_arg(args, 1) {
-        Ok(r) => heap.get(r)?.class_name.clone(),
-        Err(_) => "duke/util/ToListCollector".to_string(),
-    };
-
-    if collector_class == "duke/util/JoiningCollector" {
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let read_str_field = |heap: &duke_gc::Heap, idx: usize| -> String {
-            match heap
-                .get(collector_ref)
-                .ok()
-                .and_then(|o| o.fields.get(idx).copied())
-            {
-                Some(Slot::Reference(Some(dr))) => heap
-                    .get(dr)
-                    .ok()
-                    .and_then(|o| o.string_value.clone())
-                    .unwrap_or_default(),
-                _ => String::new(),
-            }
-        };
-        let delim = read_str_field(heap, 0);
-        let prefix = read_str_field(heap, 1);
-        let suffix = read_str_field(heap, 2);
-
-        // ⚡ Bolt: Eliminate intermediate Vec<String> allocation, format! macro overhead,
-        // and .join() by appending directly to a single String buffer.
-        // Havoc: prevent OOM from capacity overflow
-        let extra = elems.len().saturating_mul(10usize.saturating_add(delim.len()));
-        let cap = prefix.len().checked_add(suffix.len()).and_then(|x| x.checked_add(extra));
-        let max_size = 1024 * 1024 * 128; // 128 MB max string size
-
-        if cap.is_none_or(|c| c > max_size) {
-            return Err(Error::JavaException {
-                class_name: "java/lang/OutOfMemoryError".to_string(),
-            });
-        }
-
-        let mut joined = String::with_capacity(cap.unwrap());
-        joined.push_str(&prefix);
-        let mut first = true;
-        for s in &elems {
-            #[allow(clippy::collapsible_if)]
-            if let Slot::Reference(Some(r)) = s {
-                #[allow(clippy::collapsible_if)]
-                if let Ok(obj) = heap.get(*r) {
-                    if let Some(s_val) = &obj.string_value {
-                        if !first {
-                            joined.push_str(&delim);
-                        }
-                        joined.push_str(s_val);
-                        first = false;
-                    }
-                }
-            }
-        }
-        joined.push_str(&suffix);
-
-        let result_ref = heap.allocate_string(joined);
-        Ok(Some(Slot::Reference(Some(result_ref))))
-    } else if collector_class == "duke/util/CountingCollector" {
-        // collect() returns Object; box the Long so bytecode can checkcast/invokevirtual it.
-        let boxed = heap.allocate("java/lang/Long".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Long(i64::from(size));
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/GroupingByCollector" {
-        // GroupingByCollector: fields[0] = key function slot
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Ok(Some(Slot::Reference(None)));
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        // Build a HashMap: key → ArrayList of values
-        let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let key = ops
-                .invoke(
-                    heap,
-                    out,
-                    &fn_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![fn_slot, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            // find existing bucket or create new list
-            let fields = heap.get(map_ref)?.fields.clone();
-            let size_n = match fields.first() {
-                Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-                _ => 0,
-            };
-            let mut found_ki = None;
-            for i in 0..size_n {
-                let ki = 1 + i * 2;
-                if fields.get(ki).is_some_and(|k| slots_equal(k, &key, heap)) {
-                    found_ki = Some(ki);
-                    break;
-                }
-            }
-            if let Some(ki) = found_ki {
-                // Append elem to existing list
-                let list_slot = extract_field_arg(heap, map_ref, ki + 1)?;
-                if let Slot::Reference(Some(list_ref)) = list_slot {
-                    let list_size = match heap.get(list_ref)?.fields.first() {
-                        Some(Slot::Int(n)) => *n,
-                        _ => 0,
-                    };
-                    heap.get_mut(list_ref)?.fields.push(elem);
-                    heap.get_mut(list_ref)?.fields[0] = Slot::Int(list_size + 1);
-                }
-            } else {
-                // New key — create list with one element
-                let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
-                heap.get_mut(list_ref)?.fields[0] = Slot::Int(1);
-                heap.get_mut(list_ref)?.fields.push(elem);
-                heap.get_mut(map_ref)?.fields.push(key);
-                heap.get_mut(map_ref)?
-                    .fields
-                    .push(Slot::Reference(Some(list_ref)));
-                heap.get_mut(map_ref)?.fields[0] =
-                    Slot::Int(i32::try_from(size_n + 1).unwrap_or(i32::MAX));
-            }
-        }
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/ToSetCollector" {
-        // Collect into HashSet (deduplicates).
-        let set_ref = heap.allocate("java/util/HashSet".to_string(), 1);
-        heap.get_mut(set_ref)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            // Check for duplicate before inserting
-            let set_fields = heap.get(set_ref)?.fields.clone();
-            let set_size = match set_fields.first() {
-                Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-                _ => 0,
-            };
-            let already = set_fields[1..=set_size]
-                .iter()
-                .any(|s| slots_equal(s, &elem, heap));
-            if !already {
-                let cur_size = match heap.get(set_ref)?.fields.first() {
-                    Some(Slot::Int(n)) => *n,
-                    _ => 0,
-                };
-                heap.get_mut(set_ref)?.fields.push(elem);
-                heap.get_mut(set_ref)?.fields[0] = Slot::Int(cur_size + 1);
-            }
-        }
-        Ok(Some(Slot::Reference(Some(set_ref))))
-    } else if collector_class == "duke/util/ToMapCollector" {
-        // Collect into HashMap using key/val extractor functions.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let key_fn = extract_first_field_arg(heap, collector_ref)?;
-        let val_fn = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(key_ref)) = key_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let Slot::Reference(Some(val_ref)) = val_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let key_class = heap.get(key_ref)?.class_name.clone();
-        let val_class = heap.get(val_ref)?.class_name.clone();
-        let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let k_raw = ops
-                .invoke(
-                    heap,
-                    out,
-                    &key_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![key_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            let v_raw = ops
-                .invoke(
-                    heap,
-                    out,
-                    &val_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![val_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            // Box primitives so the map stores References (Object contract).
-            let k = box_primitive_slot(k_raw, heap);
-            let v = box_primitive_slot(v_raw, heap);
-            let cur_size = match heap.get(map_ref)?.fields.first() {
-                Some(Slot::Int(n)) => *n,
-                _ => 0,
-            };
-            heap.get_mut(map_ref)?.fields.push(k);
-            heap.get_mut(map_ref)?.fields.push(v);
-            heap.get_mut(map_ref)?.fields[0] = Slot::Int(cur_size + 1);
-        }
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/ToUnmodifiableMapCollector" {
-        // Same as ToMapCollector but produces an UnmodifiableMap.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let key_fn = extract_first_field_arg(heap, collector_ref)?;
-        let val_fn = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(key_ref)) = key_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let Slot::Reference(Some(val_ref)) = val_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let key_class = heap.get(key_ref)?.class_name.clone();
-        let val_class = heap.get(val_ref)?.class_name.clone();
-        let map_ref = heap.allocate("java/util/UnmodifiableMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let k_raw = ops
-                .invoke(
-                    heap,
-                    out,
-                    &key_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![key_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            let v_raw = ops
-                .invoke(
-                    heap,
-                    out,
-                    &val_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![val_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            let k = box_primitive_slot(k_raw, heap);
-            let v = box_primitive_slot(v_raw, heap);
-            let cur_size = match heap.get(map_ref)?.fields.first() {
-                Some(Slot::Int(n)) => *n,
-                _ => 0,
-            };
-            heap.get_mut(map_ref)?.fields.push(k);
-            heap.get_mut(map_ref)?.fields.push(v);
-            heap.get_mut(map_ref)?.fields[0] = Slot::Int(cur_size + 1);
-        }
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/ToMapMergeCollector" {
-        // Collect into HashMap with merge function for duplicate keys.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let key_fn = extract_first_field_arg(heap, collector_ref)?;
-        let val_fn = extract_field_arg(heap, collector_ref, 1)?;
-        let merge_fn = extract_field_arg(heap, collector_ref, 2)?;
-        let Slot::Reference(Some(key_ref)) = key_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let Slot::Reference(Some(val_ref)) = val_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let Slot::Reference(Some(merge_ref)) = merge_fn else {
-            return Err(Error::NullPointerException);
-        };
-        let key_class = heap.get(key_ref)?.class_name.clone();
-        let val_class = heap.get(val_ref)?.class_name.clone();
-        let merge_class = heap.get(merge_ref)?.class_name.clone();
-        let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let k = ops
-                .invoke(
-                    heap,
-                    out,
-                    &key_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![key_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            let v = ops
-                .invoke(
-                    heap,
-                    out,
-                    &val_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![val_fn, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            let fields = heap.get(map_ref)?.fields.clone();
-            if let Some(i) = find_hashmap_entry_index(&fields, &k, heap) {
-                // Duplicate key — apply merge function: merge(existing, new)
-                let existing = fields[i + 1];
-                let merged = ops
-                    .invoke(
-                        heap,
-                        out,
-                        &merge_class,
-                        "apply",
-                        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-                        vec![merge_fn, existing, v],
-                    )?
-                    .unwrap_or(Slot::Reference(None));
-                heap.get_mut(map_ref)?.fields[i + 1] = merged;
-            } else {
-                let cur_size = match heap.get(map_ref)?.fields.first() {
-                    Some(Slot::Int(n)) => *n,
-                    _ => 0,
-                };
-                heap.get_mut(map_ref)?.fields.push(k);
-                heap.get_mut(map_ref)?.fields.push(v);
-                heap.get_mut(map_ref)?.fields[0] = Slot::Int(cur_size + 1);
-            }
-        }
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/PartitioningByCollector" {
-        // Collect into a Map<Boolean, List> partitioned by predicate.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let pred_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(pred_ref)) = pred_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let pred_class = heap.get(pred_ref)?.class_name.clone();
-        // Create two lists and the result map.
-        let true_list = heap.allocate("java/util/ArrayList".to_string(), 1);
-        heap.get_mut(true_list)?.fields[0] = Slot::Int(0);
-        let false_list = heap.allocate("java/util/ArrayList".to_string(), 1);
-        heap.get_mut(false_list)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &pred_class,
-                "test",
-                "(Ljava/lang/Object;)Z",
-                vec![Slot::Reference(Some(pred_ref)), elem],
-            )?;
-            let is_true = matches!(result, Some(Slot::Int(n)) if n != 0);
-            let target = if is_true { true_list } else { false_list };
-            let cur_size = match heap.get(target)?.fields.first() {
-                Some(Slot::Int(n)) => *n,
-                _ => 0,
-            };
-            heap.get_mut(target)?.fields.push(elem);
-            heap.get_mut(target)?.fields[0] = Slot::Int(cur_size + 1);
-        }
-        // Build HashMap: Boolean(1)→trueList, Boolean(0)→falseList
-        let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(2);
-        let bool_true = heap.allocate("java/lang/Boolean".to_string(), 1);
-        heap.get_mut(bool_true)?.fields[0] = Slot::Int(1);
-        let bool_false = heap.allocate("java/lang/Boolean".to_string(), 1);
-        heap.get_mut(bool_false)?.fields[0] = Slot::Int(0);
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(bool_true)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(true_list)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(bool_false)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(false_list)));
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/PartitioningByDownstreamCollector" {
-        // partitioningBy(pred, downstream): partition then apply downstream to each group.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let pred_slot = extract_first_field_arg(heap, collector_ref)?;
-        let downstream_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(pred_ref)) = pred_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let pred_class = heap.get(pred_ref)?.class_name.clone();
-        let mut true_elems: Vec<Slot> = Vec::new();
-        let mut false_elems: Vec<Slot> = Vec::new();
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &pred_class,
-                "test",
-                "(Ljava/lang/Object;)Z",
-                vec![Slot::Reference(Some(pred_ref)), elem],
-            )?;
-            if matches!(result, Some(Slot::Int(n)) if n != 0) {
-                true_elems.push(elem);
-            } else {
-                false_elems.push(elem);
-            }
-        }
-        // Apply downstream collector to each partition by building a mini stream.
-        let apply_downstream = |elems_sub: Vec<Slot>,
-                                heap: &mut duke_gc::Heap,
-                                downstream: Slot|
-         -> Result<Option<Slot>> {
-            let n = elems_sub.len();
-            let downstream_class = match downstream {
-                Slot::Reference(Some(r)) => heap.get(r)?.class_name.clone(),
-                _ => return Ok(Some(Slot::Reference(None))),
-            };
-            if downstream_class == "duke/util/CountingCollector" {
-                let boxed = heap.allocate("java/lang/Long".to_string(), 1);
-                #[allow(clippy::cast_possible_wrap)] // n is a collection count; fits i64
-                let count_long = Slot::Long(n as i64);
-                heap.get_mut(boxed)?.fields[0] = count_long;
-                return Ok(Some(Slot::Reference(Some(boxed))));
-            }
-            // Generic: build a mini stream and collect into a list.
-            let stream_ref = heap.allocate("duke/util/Stream".to_string(), n);
-            #[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-            let n_i32 = n as i32;
-            heap.get_mut(stream_ref)?.fields[0] = Slot::Int(n_i32);
-            for (i, e) in elems_sub.into_iter().enumerate() {
-                heap.get_mut(stream_ref)?.fields[1 + i] = e;
-            }
-            let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
-            heap.get_mut(list_ref)?.fields[0] = Slot::Int(0);
-            for i in 1..=n {
-                let e = heap.get(stream_ref)?.fields[i];
-                let cur = match heap.get(list_ref)?.fields.first() {
-                    Some(Slot::Int(x)) => *x,
-                    _ => 0,
-                };
-                heap.get_mut(list_ref)?.fields.push(e);
-                heap.get_mut(list_ref)?.fields[0] = Slot::Int(cur + 1);
-            }
-            Ok(Some(Slot::Reference(Some(list_ref))))
-        };
-        let true_result = apply_downstream(true_elems, heap, downstream_slot)?;
-        let false_result = apply_downstream(false_elems, heap, downstream_slot)?;
-        let map_ref = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(map_ref)?.fields[0] = Slot::Int(2);
-        let bool_true = heap.allocate("java/lang/Boolean".to_string(), 1);
-        heap.get_mut(bool_true)?.fields[0] = Slot::Int(1);
-        let bool_false = heap.allocate("java/lang/Boolean".to_string(), 1);
-        heap.get_mut(bool_false)?.fields[0] = Slot::Int(0);
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(bool_true)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(true_result.unwrap_or(Slot::Reference(None)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(Slot::Reference(Some(bool_false)));
-        heap.get_mut(map_ref)?
-            .fields
-            .push(false_result.unwrap_or(Slot::Reference(None)));
-        Ok(Some(Slot::Reference(Some(map_ref))))
-    } else if collector_class == "duke/util/SummingIntCollector" {
-        // Sum via applyAsInt(elem) for each element.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0_i32;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsInt",
-                "(Ljava/lang/Object;)I",
-                vec![fn_slot, elem],
-            )?;
-            if let Some(Slot::Int(n)) = result {
-                sum = sum.wrapping_add(n);
-            }
-        }
-        // Return boxed Integer
-        let boxed = heap.allocate("java/lang/Integer".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Int(sum);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/AveragingIntCollector" {
-        // Average via applyAsInt(elem) for each element.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0_i64;
-        let mut count = 0_usize;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsInt",
-                "(Ljava/lang/Object;)I",
-                vec![fn_slot, elem],
-            )?;
-            if let Some(Slot::Int(n)) = result {
-                sum += i64::from(n);
-                count += 1;
-            }
-        }
-        #[allow(clippy::cast_precision_loss)]
-        let avg = if count == 0 {
-            0.0
-        } else {
-            sum as f64 / count as f64
-        };
-        // Return boxed Double
-        let boxed = heap.allocate("java/lang/Double".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Double(avg);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/SummarizingIntCollector" {
-        // IntSummaryStatistics via applyAsInt(elem) for each element.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0_i64;
-        let mut min = i32::MAX;
-        let mut max = i32::MIN;
-        let mut count = 0_i64;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsInt",
-                "(Ljava/lang/Object;)I",
-                vec![fn_slot, elem],
-            )?;
-            if let Some(Slot::Int(n)) = result {
-                sum += i64::from(n);
-                if n < min {
-                    min = n;
-                }
-                if n > max {
-                    max = n;
-                }
-                count += 1;
-            }
-        }
-        if count == 0 {
-            min = 0;
-            max = 0;
-        }
-        // fields: [0]=count(J) [1]=sum(J) [2]=min(I) [3]=max(I)
-        let stats = heap.allocate("java/util/IntSummaryStatistics".to_string(), 4);
-        heap.get_mut(stats)?.fields[0] = Slot::Long(count);
-        heap.get_mut(stats)?.fields[1] = Slot::Long(sum);
-        heap.get_mut(stats)?.fields[2] = Slot::Int(min);
-        heap.get_mut(stats)?.fields[3] = Slot::Int(max);
-        Ok(Some(Slot::Reference(Some(stats))))
-    } else if collector_class == "duke/util/SummingLongCollector" {
-        // Sum via applyAsLong(elem) for each element.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0_i64;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsLong",
-                "(Ljava/lang/Object;)J",
-                vec![fn_slot, elem],
-            )?;
-            match result {
-                Some(Slot::Long(n)) => sum = sum.wrapping_add(n),
-                Some(Slot::Int(n)) => sum = sum.wrapping_add(i64::from(n)),
-                _ => {}
-            }
-        }
-        // Return boxed Long
-        let boxed = heap.allocate("java/lang/Long".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Long(sum);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/AveragingDoubleCollector" {
-        // Average via applyAsDouble(elem) for each element.
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0.0_f64;
-        let mut count = 0_usize;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsDouble",
-                "(Ljava/lang/Object;)D",
-                vec![fn_slot, elem],
-            )?;
-            match result {
-                Some(Slot::Double(d)) => {
-                    sum += d;
-                    count += 1;
-                }
-                Some(Slot::Float(f)) => {
-                    sum += f64::from(f);
-                    count += 1;
-                }
-                Some(Slot::Int(n)) => {
-                    sum += f64::from(n);
-                    count += 1;
-                }
-                _ => {}
-            }
-        }
-        #[allow(clippy::cast_precision_loss)]
-        let avg = if count == 0 { 0.0 } else { sum / count as f64 };
-        // Return boxed Double
-        let boxed = heap.allocate("java/lang/Double".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Double(avg);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/MappingCollector" {
-        // MappingCollector: fields[0]=mapper fn, fields[1]=downstream collector
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let mapper_slot = extract_first_field_arg(heap, collector_ref)?;
-        let downstream_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(mapper_ref)) = mapper_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let mapper_class = heap.get(mapper_ref)?.class_name.clone();
-        // Map each element through the mapper function
-        let mut mapped_elems = Vec::with_capacity(elems.len());
-        for elem in elems {
-            let mapped = ops
-                .invoke(
-                    heap,
-                    out,
-                    &mapper_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![mapper_slot, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            mapped_elems.push(mapped);
-        }
-        // Build a temporary stream from mapped elements and collect with downstream
-        let mapped_size = i32::try_from(mapped_elems.len()).unwrap_or(0);
-        let tmp_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-        heap.get_mut(tmp_stream)?.fields[0] = Slot::Int(mapped_size);
-        for elem in mapped_elems {
-            heap.get_mut(tmp_stream)?.fields.push(elem);
-        }
-        let tmp_args = vec![Slot::Reference(Some(tmp_stream)), downstream_slot];
-        native_stream_collect(&tmp_args, heap, out, control, ops)
-    } else if collector_class == "duke/util/GroupingBy2Collector" {
-        // groupingBy(keyFn, downstream): fields[0]=keyFn, fields[1]=downstream collector
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let downstream_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        // First pass: group raw elements by key into HashMap<key, ArrayList<elem>>
-        let raw_map = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(raw_map)?.fields[0] = Slot::Int(0);
-        for elem in elems {
-            let key_raw = ops
-                .invoke(
-                    heap,
-                    out,
-                    &fn_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![fn_slot, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            // Box primitive keys so HashMap.get(boxed) can match them via slots_equal
-            let key = box_primitive_slot(key_raw, heap);
-            let fields = heap.get(raw_map)?.fields.clone();
-            let size_n = match fields.first() {
-                Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-                _ => 0,
-            };
-            let mut found_ki = None;
-            for i in 0..size_n {
-                let ki = 1 + i * 2;
-                if fields.get(ki).is_some_and(|k| slots_equal(k, &key, heap)) {
-                    found_ki = Some(ki);
-                    break;
-                }
-            }
-            if let Some(ki) = found_ki {
-                let list_slot = extract_field_arg(heap, raw_map, ki + 1)?;
-                if let Slot::Reference(Some(list_ref)) = list_slot {
-                    let list_size = match heap.get(list_ref)?.fields.first() {
-                        Some(Slot::Int(n)) => *n,
-                        _ => 0,
-                    };
-                    heap.get_mut(list_ref)?.fields.push(elem);
-                    heap.get_mut(list_ref)?.fields[0] = Slot::Int(list_size + 1);
-                }
-            } else {
-                let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
-                heap.get_mut(list_ref)?.fields[0] = Slot::Int(1);
-                heap.get_mut(list_ref)?.fields.push(elem);
-                heap.get_mut(raw_map)?.fields.push(key);
-                heap.get_mut(raw_map)?
-                    .fields
-                    .push(Slot::Reference(Some(list_ref)));
-                heap.get_mut(raw_map)?.fields[0] =
-                    Slot::Int(i32::try_from(size_n + 1).unwrap_or(i32::MAX));
-            }
-        }
-        // Second pass: apply downstream collector to each group's ArrayList
-        let result_map = heap.allocate("java/util/HashMap".to_string(), 1);
-        heap.get_mut(result_map)?.fields[0] = Slot::Int(0);
-        let raw_fields = heap.get(raw_map)?.fields.clone();
-        let group_count = match raw_fields.first() {
-            Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-            _ => 0,
-        };
-        for i in 0..group_count {
-            let key = raw_fields
-                .get(1 + i * 2)
-                .copied()
-                .unwrap_or(Slot::Reference(None));
-            let list_slot = raw_fields
-                .get(2 + i * 2)
-                .copied()
-                .unwrap_or(Slot::Reference(None));
-            let Slot::Reference(Some(list_ref)) = list_slot else {
-                continue;
-            };
-            let group_size_field = heap
-                .get(list_ref)?
-                .fields
-                .first()
-                .copied()
-                .unwrap_or(Slot::Int(0));
-            let group_size = match group_size_field {
-                Slot::Int(n) => n,
-                _ => 0,
-            };
-            let tmp_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-            heap.get_mut(tmp_stream)?.fields[0] = group_size_field;
-            let group_elems: Vec<Slot> =
-                heap.get(list_ref)?.fields[1..=usize::try_from(group_size).unwrap_or(0)].to_vec();
-            for e in group_elems {
-                heap.get_mut(tmp_stream)?.fields.push(e);
-            }
-            let tmp_args = vec![Slot::Reference(Some(tmp_stream)), downstream_slot];
-            let collected = native_stream_collect(&tmp_args, heap, out, control, ops)?
-                .unwrap_or(Slot::Reference(None));
-            let cur_result_size = match heap.get(result_map)?.fields.first() {
-                Some(Slot::Int(n)) => *n,
-                _ => 0,
-            };
-            heap.get_mut(result_map)?.fields.push(key);
-            heap.get_mut(result_map)?.fields.push(collected);
-            heap.get_mut(result_map)?.fields[0] = Slot::Int(cur_result_size + 1);
-        }
-        Ok(Some(Slot::Reference(Some(result_map))))
-    } else if collector_class == "duke/util/MinByCollector" {
-        // minBy(comparator): fields[0] = comparator
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let cmp_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(cmp_ref)) = cmp_slot else {
-            let r = make_optional(heap, None);
-            return Ok(Some(Slot::Reference(Some(r))));
-        };
-        let cmp_class = heap.get(cmp_ref)?.class_name.clone();
-        let mut min: Option<Slot> = None;
-        for elem in elems {
-            let is_less = if let Some(ref cur) = min {
-                let result = ops
-                    .invoke(
-                        heap,
-                        out,
-                        &cmp_class,
-                        "compare",
-                        "(Ljava/lang/Object;Ljava/lang/Object;)I",
-                        vec![cmp_slot, elem, *cur],
-                    )?
-                    .unwrap_or(Slot::Int(0));
-                matches!(result, Slot::Int(n) if n < 0)
-            } else {
-                true
-            };
-            if is_less {
-                min = Some(elem);
-            }
-        }
-        let r = make_optional(heap, min);
-        Ok(Some(Slot::Reference(Some(r))))
-    } else if collector_class == "duke/util/MaxByCollector" {
-        // maxBy(comparator): fields[0] = comparator
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let cmp_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(cmp_ref)) = cmp_slot else {
-            let r = make_optional(heap, None);
-            return Ok(Some(Slot::Reference(Some(r))));
-        };
-        let cmp_class = heap.get(cmp_ref)?.class_name.clone();
-        let mut max: Option<Slot> = None;
-        for elem in elems {
-            let is_greater = if let Some(ref cur) = max {
-                let result = ops
-                    .invoke(
-                        heap,
-                        out,
-                        &cmp_class,
-                        "compare",
-                        "(Ljava/lang/Object;Ljava/lang/Object;)I",
-                        vec![cmp_slot, elem, *cur],
-                    )?
-                    .unwrap_or(Slot::Int(0));
-                matches!(result, Slot::Int(n) if n > 0)
-            } else {
-                true
-            };
-            if is_greater {
-                max = Some(elem);
-            }
-        }
-        let r = make_optional(heap, max);
-        Ok(Some(Slot::Reference(Some(r))))
-    } else if collector_class == "duke/util/SummingDoubleCollector" {
-        // summingDouble: fields[0] = ToDoubleFunction
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0.0_f64;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsDouble",
-                "(Ljava/lang/Object;)D",
-                vec![fn_slot, elem],
-            )?;
-            match result {
-                Some(Slot::Double(d)) => sum += d,
-                Some(Slot::Float(f)) => sum += f64::from(f),
-                Some(Slot::Int(n)) => sum += f64::from(n),
-                _ => {}
-            }
-        }
-        let boxed = heap.allocate("java/lang/Double".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Double(sum);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/AveragingLongCollector" {
-        // averagingLong: fields[0] = ToLongFunction
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let fn_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let mut sum = 0_i64;
-        let mut count = 0_usize;
-        for elem in elems {
-            let result = ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsLong",
-                "(Ljava/lang/Object;)J",
-                vec![fn_slot, elem],
-            )?;
-            match result {
-                Some(Slot::Long(n)) => {
-                    sum = sum.wrapping_add(n);
-                    count += 1;
-                }
-                Some(Slot::Int(n)) => {
-                    sum = sum.wrapping_add(i64::from(n));
-                    count += 1;
-                }
-                _ => {}
-            }
-        }
-        #[allow(clippy::cast_precision_loss)]
-        let avg = if count == 0 {
-            0.0
-        } else {
-            sum as f64 / count as f64
-        };
-        let boxed = heap.allocate("java/lang/Double".to_string(), 1);
-        heap.get_mut(boxed)?.fields[0] = Slot::Double(avg);
-        Ok(Some(Slot::Reference(Some(boxed))))
-    } else if collector_class == "duke/util/ReducingNoIdentityCollector" {
-        // reducing(BinaryOperator) → Optional<T>
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let op_slot = extract_first_field_arg(heap, collector_ref)?;
-        let Slot::Reference(Some(bop_ref)) = op_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let op_class = heap.get(bop_ref)?.class_name.clone();
-        let result = if elems.is_empty() {
-            None
-        } else {
-            let mut acc = elems[0];
-            for elem in elems.into_iter().skip(1) {
-                acc = ops
-                    .invoke(
-                        heap,
-                        out,
-                        &op_class,
-                        "apply",
-                        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-                        vec![op_slot, acc, elem],
-                    )?
-                    .unwrap_or(Slot::Reference(None));
-            }
-            Some(acc)
-        };
-        let opt_ref = make_optional(heap, result);
-        Ok(Some(Slot::Reference(Some(opt_ref))))
-    } else if collector_class == "duke/util/ReducingCollector" {
-        // reducing(identity, BinaryOperator) → T
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let identity_slot = extract_first_field_arg(heap, collector_ref)?;
-        let op_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(bop_ref)) = op_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let op_class = heap.get(bop_ref)?.class_name.clone();
-        let mut acc = identity_slot;
-        for elem in elems {
-            acc = ops
-                .invoke(
-                    heap,
-                    out,
-                    &op_class,
-                    "apply",
-                    "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![op_slot, acc, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-        }
-        Ok(Some(acc))
-    } else if collector_class == "duke/util/ReducingMappingCollector" {
-        // reducing(identity, mapper, BinaryOperator) → U
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let identity_slot = extract_first_field_arg(heap, collector_ref)?;
-        let mapper_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let op_slot = extract_field_arg(heap, collector_ref, 2)?;
-        let Slot::Reference(Some(mapper_ref)) = mapper_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let Slot::Reference(Some(bop_ref)) = op_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let mapper_class = heap.get(mapper_ref)?.class_name.clone();
-        let op_class = heap.get(bop_ref)?.class_name.clone();
-        let mut acc = identity_slot;
-        for elem in elems {
-            let mapped = ops
-                .invoke(
-                    heap,
-                    out,
-                    &mapper_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![mapper_slot, elem],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            acc = ops
-                .invoke(
-                    heap,
-                    out,
-                    &op_class,
-                    "apply",
-                    "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![op_slot, acc, mapped],
-                )?
-                .unwrap_or(Slot::Reference(None));
-        }
-        Ok(Some(acc))
-    } else if collector_class == "duke/util/CollectingAndThenCollector" {
-        // collectingAndThen(downstream, finisher): fields[0]=downstream, fields[1]=finisher
-        let collector_ref = extract_ref_arg(args, 1)?;
-        let downstream_slot = extract_first_field_arg(heap, collector_ref)?;
-        let finisher_slot = extract_field_arg(heap, collector_ref, 1)?;
-        let Slot::Reference(Some(finisher_ref)) = finisher_slot else {
-            return Err(Error::NullPointerException);
-        };
-        let finisher_class = heap.get(finisher_ref)?.class_name.clone();
-        // First collect with downstream
-        let tmp_args = vec![Slot::Reference(Some(stream_ref)), downstream_slot];
-        let intermediate = native_stream_collect(&tmp_args, heap, out, control, ops)?
-            .unwrap_or(Slot::Reference(None));
-        // Then apply finisher
-        let result = ops.invoke(
-            heap,
-            out,
-            &finisher_class,
-            "apply",
-            "(Ljava/lang/Object;)Ljava/lang/Object;",
-            vec![finisher_slot, intermediate],
-        )?;
-        Ok(result.or(Some(Slot::Reference(None))))
-    } else if collector_class == "duke/util/ToUnmodifiableListCollector" {
-        // toUnmodifiableList(): collect into UnmodifiableList (mutations throw).
-        let list_ref = heap.allocate("java/util/UnmodifiableList".to_string(), 1);
-        heap.get_mut(list_ref)?.fields[0] = Slot::Int(size);
-        for elem in elems {
-            heap.get_mut(list_ref)?.fields.push(elem);
-        }
-        Ok(Some(Slot::Reference(Some(list_ref))))
-    } else {
-        // ToListCollector (default): collect into ArrayList.
-        let list_ref = heap.allocate("java/util/ArrayList".to_string(), 1);
-        heap.get_mut(list_ref)?.fields[0] = Slot::Int(size);
-        for elem in elems {
-            heap.get_mut(list_ref)?.fields.push(elem);
-        }
-        Ok(Some(Slot::Reference(Some(list_ref))))
-    }
-}
-
-/// Native: `Stream.distinct()Stream` — removes duplicate elements (by `slots_equal`).
-pub(crate) fn native_stream_distinct(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let mut seen: Vec<Slot> = Vec::new();
-    for elem in elems {
-        if !seen.iter().any(|s| slots_equal(s, &elem, heap)) {
-            seen.push(elem);
-        }
-    }
-    let new_size = i32::try_from(seen.len()).unwrap_or(0);
-    let new_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(new_stream)?.fields[0] = Slot::Int(new_size);
-    for elem in seen {
-        heap.get_mut(new_stream)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(new_stream))))
-}
-
-/// Native: `Stream.sorted()Stream` — sorts elements by natural order via `compareTo`.
-pub(crate) fn native_stream_sorted(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let mut elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    // Insertion sort via compareTo callbacks (stable, O(n²) — fine for test sizes).
-    for i in 1..elems.len() {
-        let mut j = i;
-        while j > 0 {
-            let Slot::Reference(Some(a)) = elems[j - 1] else {
-                break;
-            };
-            let Slot::Reference(Some(b)) = elems[j] else {
-                break;
-            };
-            let class_a = heap.get(a)?.class_name.clone();
-            let cmp = ops.invoke(
-                heap,
-                out,
-                &class_a,
-                "compareTo",
-                "(Ljava/lang/Object;)I",
-                vec![Slot::Reference(Some(a)), Slot::Reference(Some(b))],
-            )?;
-            if matches!(cmp, Some(Slot::Int(n)) if n > 0) {
-                elems.swap(j - 1, j);
-                j -= 1;
-            } else {
-                break;
-            }
-        }
-    }
-    let new_size = i32::try_from(elems.len()).unwrap_or(0);
-    let new_stream = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(new_stream)?.fields[0] = Slot::Int(new_size);
-    for elem in elems {
-        heap.get_mut(new_stream)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(new_stream))))
-}
-
-/// Native: `Stream.anyMatch(Predicate)Z` — true if any element satisfies predicate.
-pub(crate) fn native_stream_any_match(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let pred_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(pred_ref)) = pred_slot else {
-        return Ok(Some(Slot::Int(0)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let pred_class = heap.get(pred_ref)?.class_name.clone();
-    for elem in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &pred_class,
-            "test",
-            "(Ljava/lang/Object;)Z",
-            vec![Slot::Reference(Some(pred_ref)), elem],
-        )?;
-        if matches!(result, Some(Slot::Int(1))) {
-            return Ok(Some(Slot::Int(1)));
-        }
-    }
-    Ok(Some(Slot::Int(0)))
-}
-
-/// Native: `Stream.allMatch(Predicate)Z` — true if all elements satisfy predicate.
-pub(crate) fn native_stream_all_match(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let pred_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(pred_ref)) = pred_slot else {
-        return Ok(Some(Slot::Int(1)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let pred_class = heap.get(pred_ref)?.class_name.clone();
-    for elem in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &pred_class,
-            "test",
-            "(Ljava/lang/Object;)Z",
-            vec![Slot::Reference(Some(pred_ref)), elem],
-        )?;
-        if !matches!(result, Some(Slot::Int(1))) {
-            return Ok(Some(Slot::Int(0)));
-        }
-    }
-    Ok(Some(Slot::Int(1)))
-}
-
-/// Native: `Stream.noneMatch(Predicate)Z` — true if no element satisfies predicate.
-pub(crate) fn native_stream_none_match(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let pred_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(pred_ref)) = pred_slot else {
-        return Ok(Some(Slot::Int(1)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let pred_class = heap.get(pred_ref)?.class_name.clone();
-    for elem in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &pred_class,
-            "test",
-            "(Ljava/lang/Object;)Z",
-            vec![Slot::Reference(Some(pred_ref)), elem],
-        )?;
-        if matches!(result, Some(Slot::Int(1))) {
-            return Ok(Some(Slot::Int(0)));
-        }
-    }
-    Ok(Some(Slot::Int(1)))
-}
-
-/// Native: `Stream.findFirst()Optional` — returns Optional of first element, or empty.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_stream_find_first(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let opt_ref = heap.allocate("java/util/Optional".to_string(), 1);
-    if size > 0 {
-        let first = heap.get(stream_ref)?.fields[1];
-        heap.get_mut(opt_ref)?.fields[0] = first;
-    } else {
-        heap.get_mut(opt_ref)?.fields[0] = Slot::Reference(None);
-    }
-    Ok(Some(Slot::Reference(Some(opt_ref))))
-}
-
-/// Native: `Stream.reduce(BinaryOperator)Optional` — folds elements left via binary op.
-pub(crate) fn native_stream_reduce(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let op_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(op_ref)) = op_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let op_class = heap.get(op_ref)?.class_name.clone();
-    let reduce_result_ref = heap.allocate("java/util/Optional".to_string(), 1);
-    if elems.is_empty() {
-        heap.get_mut(reduce_result_ref)?.fields[0] = Slot::Reference(None);
-        return Ok(Some(Slot::Reference(Some(reduce_result_ref))));
-    }
-    let mut acc = elems[0];
-    for elem in elems.into_iter().skip(1) {
-        let result = ops.invoke(
-            heap,
-            out,
-            &op_class,
-            "apply",
-            "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-            vec![Slot::Reference(Some(op_ref)), acc, elem],
-        )?;
-        acc = result.unwrap_or(Slot::Reference(None));
-    }
-    // Box primitive accumulator before storing in Optional (Java generics always hold References)
-    let acc_boxed = box_primitive_slot(acc, heap);
-    heap.get_mut(reduce_result_ref)?.fields[0] = acc_boxed;
-    Ok(Some(Slot::Reference(Some(reduce_result_ref))))
-}
-
-/// Native: `Stream.reduce(identity, BinaryOperator)Object` — fold with initial value.
-pub(crate) fn native_stream_reduce_with_identity(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let identity = extract_slot_arg(args, 1);
-    let fn_slot = extract_slot_arg(args, 2);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(identity));
-    };
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let mut acc = identity;
-    for elem in elems {
-        acc = ops
-            .invoke(
-                heap,
-                out,
-                &fn_class,
-                "apply",
-                "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-                vec![fn_slot, acc, elem],
-            )?
-            .unwrap_or(Slot::Reference(None));
-    }
-    // Autobox: if the identity was a Reference (stream of boxed type) but the accumulator
-    // impl returned a raw primitive (e.g. Integer::sum returns int), re-box the result so
-    // that the caller can apply intValue() / longValue() as expected.
-    let acc = match (identity, acc) {
-        (Slot::Reference(_), Slot::Int(v)) => {
-            let r = heap.allocate("java/lang/Integer".to_string(), 1);
-            heap.get_mut(r)?.fields[0] = Slot::Int(v);
-            Slot::Reference(Some(r))
-        }
-        (Slot::Reference(_), Slot::Long(v)) => {
-            let r = heap.allocate("java/lang/Long".to_string(), 1);
-            heap.get_mut(r)?.fields[0] = Slot::Long(v);
-            Slot::Reference(Some(r))
-        }
-        (Slot::Reference(_), Slot::Double(v)) => {
-            let r = heap.allocate("java/lang/Double".to_string(), 1);
-            heap.get_mut(r)?.fields[0] = Slot::Double(v);
-            Slot::Reference(Some(r))
-        }
-        _ => acc,
-    };
-    Ok(Some(acc))
-}
-
-/// Native: `Stream.toList()List` — terminal op returning an unmodifiable list (same as collect).
-pub(crate) fn native_stream_to_list(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    native_stream_collect(args, heap, out, control, ops)
-}
-
-/// Helper: allocate a `JoiningCollector` with 3 fields: delimiter, prefix, suffix.
-fn make_joining_collector(
-    heap: &mut duke_gc::Heap,
-    delimiter: &str,
-    prefix: &str,
-    suffix: &str,
-) -> u64 {
-    let collector_ref = heap.allocate("duke/util/JoiningCollector".to_string(), 3);
-    let delim_ref = heap.allocate_string(delimiter.to_string());
-    let prefix_ref = heap.allocate_string(prefix.to_string());
-    let suffix_ref = heap.allocate_string(suffix.to_string());
-    let obj = heap.get_mut(collector_ref).expect("just allocated");
-    obj.fields[0] = Slot::Reference(Some(delim_ref));
-    obj.fields[1] = Slot::Reference(Some(prefix_ref));
-    obj.fields[2] = Slot::Reference(Some(suffix_ref));
-    collector_ref
-}
-
-/// Native: `Collectors.joining(delim)Collector` — returns a joining collector with delimiter.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_joining(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let delim = match args.first() {
-        Some(Slot::Reference(Some(r))) => heap
-            .get(*r)
-            .ok()
-            .and_then(|o| o.string_value.clone())
-            .unwrap_or_default(),
-        _ => String::new(),
-    };
-    let collector_ref = make_joining_collector(heap, &delim, "", "");
-    Ok(Some(Slot::Reference(Some(collector_ref))))
-}
-
-/// Native: `Collectors.joining()Collector` — no-arg version (empty delimiter).
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_joining_no_arg(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let collector_ref = make_joining_collector(heap, "", "", "");
-    Ok(Some(Slot::Reference(Some(collector_ref))))
-}
-
-/// Native: `Collectors.joining(delim, prefix, suffix)Collector` — full 3-arg joining collector.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_joining_full(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let read_str = |heap: &duke_gc::Heap, idx: usize| -> String {
-        match args.get(idx) {
-            Some(Slot::Reference(Some(r))) => heap
-                .get(*r)
-                .ok()
-                .and_then(|o| o.string_value.clone())
-                .unwrap_or_default(),
-            _ => String::new(),
-        }
-    };
-    let delim = read_str(heap, 0);
-    let prefix = read_str(heap, 1);
-    let suffix = read_str(heap, 2);
-    let collector_ref = make_joining_collector(heap, &delim, &prefix, &suffix);
-    Ok(Some(Slot::Reference(Some(collector_ref))))
-}
-
-/// Native: `Collectors.toList()Collector` — returns a sentinel collector object.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_to_list(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let collector_ref = heap.allocate("duke/util/ToListCollector".to_string(), 0);
-    Ok(Some(Slot::Reference(Some(collector_ref))))
-}
-
-/// Native: `Collectors.counting()Collector` — returns a counting collector sentinel.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_counting(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = heap.allocate("duke/util/CountingCollector".to_string(), 0);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Collectors.groupingBy(Function)Collector` — returns a grouping-by collector.
-/// Stores `fn_slot` in `fields[0]`.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_collectors_grouping_by(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let fn_slot = extract_slot_arg(args, 0);
-    let r = heap.allocate("duke/util/GroupingByCollector".to_string(), 1);
-    heap.get_mut(r)?.fields[0] = fn_slot;
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Stream.peek(Consumer)Stream` — side-effect each element, returns same stream.
-pub(crate) fn native_stream_peek(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let consumer_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(consumer_ref)) = consumer_slot else {
-        return Ok(Some(Slot::Reference(Some(stream_ref))));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let consumer_class = heap.get(consumer_ref)?.class_name.clone();
-    for elem in &elems {
-        ops.invoke(
-            heap,
-            out,
-            &consumer_class,
-            "accept",
-            "(Ljava/lang/Object;)V",
-            vec![consumer_slot, *elem],
-        )?;
-    }
-    // Return a new stream with same elements (consumer may have GC'd things)
-    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(out_ref)?.fields[0] = Slot::Int(i32::try_from(elems.len()).unwrap_or(0));
-    for elem in elems {
-        heap.get_mut(out_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(out_ref))))
-}
-
-/// Native: `Stream.toArray()Object[]` — materializes stream into an Object array.
-pub(crate) fn native_stream_to_array(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    // Arrays use fields directly (no length header); arraylength returns fields.len().
-    let arr_ref = heap.allocate("[Ljava/lang/Object;".to_string(), 0);
-    for elem in elems {
-        heap.get_mut(arr_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-// ---------------------------------------------------------------------------
-// Stream.limit / Stream.skip / Stream.flatMap
-// ---------------------------------------------------------------------------
-
-/// Native: `Stream.limit(long)Stream` — keeps first N elements.
-pub(crate) fn native_stream_limit(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let max_size = match args.get(1).copied() {
-        Some(Slot::Long(n)) => usize::try_from(n.max(0)).unwrap_or(0),
-        Some(Slot::Int(n)) => usize::try_from(n.max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    let class_name = heap.get(stream_ref)?.class_name.clone();
-
-    // Lazy generators: materialise N elements on limit().
-    if class_name == "duke/util/GeneratorStream" {
-        let supplier_slot = extract_first_field_arg(heap, stream_ref)?;
-        let Slot::Reference(Some(sup_ref)) = supplier_slot else {
-            return Ok(Some(Slot::Reference(None)));
-        };
-        let sup_class = heap.get(sup_ref)?.class_name.clone();
-        let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-        heap.get_mut(out_ref)?.fields[0] = Slot::Int(i32::try_from(max_size).unwrap_or(0));
-        for _ in 0..max_size {
-            let elem = ops
-                .invoke(
-                    heap,
-                    out,
-                    &sup_class,
-                    "get",
-                    "()Ljava/lang/Object;",
-                    vec![Slot::Reference(Some(sup_ref))],
-                )?
-                .unwrap_or(Slot::Reference(None));
-            heap.get_mut(out_ref)?.fields.push(elem);
-        }
-        return Ok(Some(Slot::Reference(Some(out_ref))));
-    }
-
-    if class_name == "duke/util/IteratorStream" {
-        // fields[0] = current seed, fields[1] = UnaryOperator fn
-        let seed = extract_first_field_arg(heap, stream_ref)?;
-        let fn_slot = extract_field_arg(heap, stream_ref, 1)?;
-        let Slot::Reference(Some(fn_ref)) = fn_slot else {
-            return Ok(Some(Slot::Reference(None)));
-        };
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-        heap.get_mut(out_ref)?.fields[0] = Slot::Int(i32::try_from(max_size).unwrap_or(0));
-        let mut current = seed;
-        for _ in 0..max_size {
-            heap.get_mut(out_ref)?.fields.push(current);
-            current = ops
-                .invoke(
-                    heap,
-                    out,
-                    &fn_class,
-                    "apply",
-                    "(Ljava/lang/Object;)Ljava/lang/Object;",
-                    vec![fn_slot, current],
-                )?
-                .unwrap_or(Slot::Reference(None));
-        }
-        return Ok(Some(Slot::Reference(Some(out_ref))));
-    }
-
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let take = size.min(max_size);
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(out_ref)?.fields[0] = Slot::Int(i32::try_from(take).unwrap_or(0));
-    for elem in elems.into_iter().take(take) {
-        heap.get_mut(out_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(out_ref))))
-}
-
-/// Native: `Stream.skip(long)Stream` — skips first N elements.
-pub(crate) fn native_stream_skip(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let skip_n = match args.get(1).copied() {
-        Some(Slot::Long(n)) => usize::try_from(n.max(0)).unwrap_or(0),
-        Some(Slot::Int(n)) => usize::try_from(n.max(0)).unwrap_or(0),
-        _ => 0,
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let skipped: Vec<Slot> = elems.into_iter().skip(skip_n).collect();
-    let new_size = i32::try_from(skipped.len()).unwrap_or(0);
-    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(out_ref)?.fields[0] = Slot::Int(new_size);
-    for elem in skipped {
-        heap.get_mut(out_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(out_ref))))
-}
-
-/// Native: `Stream.flatMap(Function)Stream` — maps each element to a Stream and flattens.
-pub(crate) fn native_stream_flat_map(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let stream_ref = extract_ref_arg(args, 0)?;
-    let fn_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let size = match heap.get(stream_ref)?.fields.first() {
-        Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-        _ => 0,
-    };
-    let elems: Vec<Slot> = heap.get(stream_ref)?.fields[1..=size].to_vec();
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let mut flat: Vec<Slot> = Vec::with_capacity(elems.len());
-    for elem in elems {
-        let inner = ops.invoke(
-            heap,
-            out,
-            &fn_class,
-            "apply",
-            "(Ljava/lang/Object;)Ljava/lang/Object;",
-            vec![fn_slot, elem],
-        )?;
-        if let Some(Slot::Reference(Some(inner_ref))) = inner {
-            // inner should be a duke/util/Stream — flatten its elements
-            let inner_size = match heap.get(inner_ref)?.fields.first() {
-                Some(Slot::Int(n)) => usize::try_from(*n).unwrap_or(0),
-                _ => 0,
-            };
-            let inner_elems: Vec<Slot> = heap.get(inner_ref)?.fields[1..=inner_size].to_vec();
-            flat.extend(inner_elems);
-        }
-    }
-    let new_size = i32::try_from(flat.len()).unwrap_or(0);
-    let out_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(out_ref)?.fields[0] = Slot::Int(new_size);
-    for elem in flat {
-        heap.get_mut(out_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(out_ref))))
-}
-
-// ---------------------------------------------------------------------------
-// IntStream — duke/util/IntStream
-// fields[0]=Int(size), fields[1..n]=Int(value) elements (unboxed ints)
-// ---------------------------------------------------------------------------
-
-/// Build an `IntStream` heap object from a `Vec<i32>`.
-fn make_int_stream(heap: &mut duke_gc::Heap, values: Vec<i32>) -> u64 {
-    let n = i32::try_from(values.len()).unwrap_or(0);
-    let r = heap.allocate("duke/util/IntStream".to_string(), 1);
-    heap.get_mut(r).expect("fresh").fields[0] = Slot::Int(n);
-    for v in values {
-        heap.get_mut(r).expect("fresh").fields.push(Slot::Int(v));
-    }
-    r
-}
-
-/// Create an `OptionalInt` heap object. `None` = empty, `Some(v)` = present.
-fn make_optional_int(heap: &mut duke_gc::Heap, value: Option<i32>) -> u64 {
-    let r = heap.allocate("duke/util/OptionalInt".to_string(), 2);
-    if let Some(v) = value {
-        heap.get_mut(r).expect("fresh").fields[0] = Slot::Int(v);
-        heap.get_mut(r).expect("fresh").fields[1] = Slot::Int(1);
-    } else {
-        heap.get_mut(r).expect("fresh").fields[1] = Slot::Int(0);
-    }
-    r
-}
-
-/// Create a `java/util/Optional` heap object. `None` = empty, `Some(slot)` = present.
-fn make_optional(heap: &mut duke_gc::Heap, value: Option<Slot>) -> u64 {
-    let r = heap.allocate("java/util/Optional".to_string(), 1);
-    heap.get_mut(r).expect("fresh").fields[0] = value.unwrap_or(Slot::Reference(None));
-    r
-}
-
-/// Extract int elements from an `IntStream` heap object.
-fn int_stream_elems(heap: &duke_gc::Heap, ref_: u64) -> Vec<i32> {
-    let size = match heap.get(ref_).ok().and_then(|o| o.fields.first().copied()) {
-        Some(Slot::Int(n)) => usize::try_from(n).unwrap_or(0),
-        _ => 0,
-    };
-    heap.get(ref_)
-        .ok()
-        .map(|o| {
-            o.fields[1..=size]
-                .iter()
-                .filter_map(|s| if let Slot::Int(n) = s { Some(*n) } else { None })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-/// Native: `IntStream.range(int,int)IntStream` — half-open range [start, end).
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_range(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let start = match args.first().copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    let end = match args.get(1).copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    let values: Vec<i32> = (start..end).collect();
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
-}
-
-/// Native: `IntStream.rangeClosed(int,int)IntStream` — inclusive range [start, end].
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_range_closed(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let start = match args.first().copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    let end = match args.get(1).copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    let values: Vec<i32> = (start..=end).collect();
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
-}
-
-/// Native: `IntStream.of(int...)IntStream` — from an int array argument.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_of(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    // args[0] is the int[] array ref
-    let arr_ref = extract_ref_arg(args, 0)?;
-    let values: Vec<i32> = heap
-        .get(arr_ref)?
-        .fields
-        .iter()
-        .filter_map(|s| if let Slot::Int(n) = s { Some(*n) } else { None })
-        .collect();
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
-}
-
-/// Native: `IntStream.iterate(seed, UnaryOperator)IntStream` — generates up to 4096 elements.
-pub(crate) fn native_int_stream_iterate(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    const MAX: usize = 4096;
-    let seed = match args.first().copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    let fn_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(Slot::Reference(Some(make_int_stream(
-            heap,
-            vec![seed],
-        )))));
-    };
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let mut values = Vec::with_capacity(MAX);
-    let mut cur = seed;
-    for _ in 0..MAX {
-        values.push(cur);
-        let next = ops
-            .invoke(
-                heap,
-                out,
-                &fn_class,
-                "applyAsInt",
-                "(I)I",
-                vec![fn_slot, Slot::Int(cur)],
-            )?
-            .unwrap_or(Slot::Int(cur));
-        match next {
-            Slot::Int(n) => cur = n,
-            _ => break,
-        }
-    }
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, values)))))
-}
-
-/// Native: `IntStream.count()J`
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_count(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let n = match heap.get(r)?.fields.first() {
-        Some(Slot::Int(n)) => i64::from(*n),
-        _ => 0,
-    };
-    Ok(Some(Slot::Long(n)))
-}
-
-/// Native: `IntStream.sum()I`
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_sum(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let sum: i32 = int_stream_elems(heap, r)
-        .iter()
-        .copied()
-        .fold(0_i32, i32::wrapping_add);
-    Ok(Some(Slot::Int(sum)))
-}
-
-/// Native: `IntStream.min()OptionalInt`
-pub(crate) fn native_int_stream_min(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    // OptionalInt: fields[0]=Int(value), fields[1]=Int(1=present/0=empty)
-    let opt_ref = heap.allocate("duke/util/OptionalInt".to_string(), 2);
-    if let Some(&v) = elems.iter().min() {
-        heap.get_mut(opt_ref)?.fields[0] = Slot::Int(v);
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(1);
-    } else {
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(0);
-    }
-    Ok(Some(Slot::Reference(Some(opt_ref))))
-}
-
-/// Native: `IntStream.max()OptionalInt`
-pub(crate) fn native_int_stream_max(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    let opt_ref = heap.allocate("duke/util/OptionalInt".to_string(), 2);
-    if let Some(&v) = elems.iter().max() {
-        heap.get_mut(opt_ref)?.fields[0] = Slot::Int(v);
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(1);
-    } else {
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(0);
-    }
-    Ok(Some(Slot::Reference(Some(opt_ref))))
-}
-
-/// Native: `IntStream.average()OptionalDouble`
-pub(crate) fn native_int_stream_average(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    // OptionalDouble: fields[0]=Double(value), fields[1]=Int(1=present/0=empty)
-    let opt_ref = heap.allocate("duke/util/OptionalDouble".to_string(), 2);
-    if elems.is_empty() {
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(0);
-    } else {
-        let sum: i64 = elems.iter().map(|&n| i64::from(n)).sum();
-        #[allow(clippy::cast_precision_loss)]
-        let avg = sum as f64 / elems.len() as f64;
-        heap.get_mut(opt_ref)?.fields[0] = Slot::Double(avg);
-        heap.get_mut(opt_ref)?.fields[1] = Slot::Int(1);
-    }
-    Ok(Some(Slot::Reference(Some(opt_ref))))
-}
-
-/// Native: `IntStream.toArray()int[]`
-pub(crate) fn native_int_stream_to_array(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    let arr_ref = heap.allocate("[I".to_string(), elems.len());
-    for (i, v) in elems.into_iter().enumerate() {
-        heap.get_mut(arr_ref)?.fields[i] = Slot::Int(v);
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-/// Native: `IntStream.filter(IntPredicate)IntStream`
-pub(crate) fn native_int_stream_filter(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let pred_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(pred_ref)) = pred_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let elems = int_stream_elems(heap, r);
-    let pred_class = heap.get(pred_ref)?.class_name.clone();
-    let mut kept = Vec::with_capacity(elems.len());
-    for v in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &pred_class,
-            "test",
-            "(I)Z",
-            vec![pred_slot, Slot::Int(v)],
-        )?;
-        if matches!(result, Some(Slot::Int(n)) if n != 0) {
-            kept.push(v);
-        }
-    }
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, kept)))))
-}
-
-/// Native: `IntStream.peek(IntConsumer)IntStream` — calls consumer for each element, returns same stream.
-pub(crate) fn native_int_stream_peek(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let fn_slot = extract_slot_arg(args, 1);
-    let elems = int_stream_elems(heap, r);
-    if let Slot::Reference(Some(fn_ref)) = fn_slot {
-        let fn_class = heap.get(fn_ref)?.class_name.clone();
-        for &v in &elems {
-            ops.invoke(
-                heap,
-                out,
-                &fn_class,
-                "accept",
-                "(I)V",
-                vec![fn_slot, Slot::Int(v)],
-            )?;
-        }
-    }
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, elems)))))
-}
-
-/// Native: `IntStream.map(IntUnaryOperator)IntStream`
-pub(crate) fn native_int_stream_map(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let fn_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let elems = int_stream_elems(heap, r);
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let mut result = Vec::with_capacity(elems.len());
-    for v in elems {
-        let r = ops.invoke(
-            heap,
-            out,
-            &fn_class,
-            "applyAsInt",
-            "(I)I",
-            vec![fn_slot, Slot::Int(v)],
-        )?;
-        result.push(match r {
-            Some(Slot::Int(n)) => n,
-            _ => 0,
-        });
-    }
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, result)))))
-}
-
-/// Native: `IntStream.forEach(IntConsumer)V`
-pub(crate) fn native_int_stream_for_each(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let consumer_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(consumer_ref)) = consumer_slot else {
-        return Ok(None);
-    };
-    let elems = int_stream_elems(heap, r);
-    let consumer_class = heap.get(consumer_ref)?.class_name.clone();
-    for v in elems {
-        ops.invoke(
-            heap,
-            out,
-            &consumer_class,
-            "accept",
-            "(I)V",
-            vec![consumer_slot, Slot::Int(v)],
-        )?;
-    }
-    Ok(None)
-}
-
-/// Native: `IntStream.boxed()Stream` — wraps each int into `java/lang/Integer`.
-pub(crate) fn native_int_stream_boxed(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    let n = i32::try_from(elems.len()).unwrap_or(0);
-    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(n);
-    for v in elems {
-        let boxed_ref = heap.allocate("java/lang/Integer".to_string(), 1);
-        heap.get_mut(boxed_ref)?.fields[0] = Slot::Int(v);
-        heap.get_mut(stream_ref)?
-            .fields
-            .push(Slot::Reference(Some(boxed_ref)));
-    }
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `IntStream.mapToObj(IntFunction)Stream` — maps ints to objects.
-pub(crate) fn native_int_stream_map_to_obj(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let fn_slot = extract_slot_arg(args, 1);
-    let Slot::Reference(Some(fn_ref)) = fn_slot else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    let elems = int_stream_elems(heap, r);
-    let fn_class = heap.get(fn_ref)?.class_name.clone();
-    let mut mapped = Vec::new();
-    for v in elems {
-        let result = ops.invoke(
-            heap,
-            out,
-            &fn_class,
-            "apply",
-            "(I)Ljava/lang/Object;",
-            vec![fn_slot, Slot::Int(v)],
-        )?;
-        mapped.push(result.unwrap_or(Slot::Reference(None)));
-    }
-    let new_size = i32::try_from(mapped.len()).unwrap_or(0);
-    let stream_ref = heap.allocate("duke/util/Stream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(new_size);
-    for elem in mapped {
-        heap.get_mut(stream_ref)?.fields.push(elem);
-    }
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `IntStream.distinct()IntStream` — removes duplicate int values.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_int_stream_distinct(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let elems = int_stream_elems(heap, r);
-    let mut seen: Vec<i32> = Vec::new();
-    for v in elems {
-        if !seen.contains(&v) {
-            seen.push(v);
-        }
-    }
-    Ok(Some(Slot::Reference(Some(make_int_stream(heap, seen)))))
-}
-
-/// Native: `String.<init>(String)V` — copy constructor: copies `string_value` from source.
-pub(crate) fn native_string_init_copy(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let this_ref = extract_ref_arg(args, 0)?;
-    let src_val = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap.get(*r)?.string_value.clone(),
-        _ => None,
-    };
-    heap.get_mut(this_ref)?.string_value = src_val;
-    Ok(None)
-}
-
-/// Native: `OptionalInt.getAsInt()I`
-pub(crate) fn native_optional_int_get_as_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let v = match heap.get(r)?.fields.first().copied() {
-        Some(Slot::Int(n)) => n,
-        _ => 0,
-    };
-    Ok(Some(Slot::Int(v)))
-}
-
-/// Native: `OptionalInt.isPresent()Z`
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_optional_int_is_present(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let present = matches!(heap.get(r)?.fields.get(1), Some(Slot::Int(1)));
-    Ok(Some(Slot::Int(i32::from(present))))
-}
-
-/// Native: `OptionalInt.orElse(int)I` — returns value if present, else the default.
-pub(crate) fn native_optional_int_or_else(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let present = matches!(heap.get(r)?.fields.get(1), Some(Slot::Int(1)));
-    if present {
-        Ok(Some(
-            heap.get(r)?.fields.first().copied().unwrap_or(Slot::Int(0)),
-        ))
-    } else {
-        Ok(Some(args.get(1).copied().unwrap_or(Slot::Int(0))))
-    }
-}
-
-/// Native: `OptionalInt.of(int)OptionalInt` — creates a present `OptionalInt`.
-pub(crate) fn native_optional_int_of(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let value = args.first().copied().unwrap_or(Slot::Int(0));
-    let r = heap.allocate("duke/util/OptionalInt".to_string(), 2);
-    heap.get_mut(r)?.fields[0] = value;
-    heap.get_mut(r)?.fields[1] = Slot::Int(1); // present = true
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `OptionalInt.empty()OptionalInt` — creates an empty `OptionalInt`.
-#[allow(clippy::unnecessary_wraps)]
-pub(crate) fn native_optional_int_empty(
-    _args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = heap.allocate("duke/util/OptionalInt".to_string(), 2);
-    heap.get_mut(r)?.fields[0] = Slot::Int(0);
-    heap.get_mut(r)?.fields[1] = Slot::Int(0); // present = false
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `OptionalLong.orElse(long)J`
-pub(crate) fn native_optional_long_or_else(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let present = matches!(heap.get(r)?.fields.get(1), Some(Slot::Int(1)));
-    if present {
-        Ok(Some(
-            heap.get(r)?
-                .fields
-                .first()
-                .copied()
-                .unwrap_or(Slot::Long(0)),
-        ))
-    } else {
-        Ok(Some(args.get(1).copied().unwrap_or(Slot::Long(0))))
-    }
-}
-
-/// Native: `OptionalDouble.orElse(double)D`
-pub(crate) fn native_optional_double_or_else(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let present = matches!(heap.get(r)?.fields.get(1), Some(Slot::Int(1)));
-    if present {
-        Ok(Some(
-            heap.get(r)?
-                .fields
-                .first()
-                .copied()
-                .unwrap_or(Slot::Double(0.0)),
-        ))
-    } else {
-        Ok(Some(args.get(1).copied().unwrap_or(Slot::Double(0.0))))
-    }
-}
-
-/// Native: `OptionalDouble.getAsDouble()D`
-pub(crate) fn native_optional_double_get_as_double(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let r = extract_ref_arg(args, 0)?;
-    let v = match heap.get(r)?.fields.first().copied() {
-        Some(Slot::Double(d)) => d,
-        _ => 0.0,
-    };
-    Ok(Some(Slot::Double(v)))
-}
-
+include!("native/stream.rs");
 // ---- ArrayDeque natives ----
 // fields[0]=Int(size), fields[1..size]=elements (front at index 1)
 
@@ -16619,425 +12570,7 @@ pub(crate) fn native_string_tostring(
     Ok(Some(extract_slot_arg(args, 0)))
 }
 
-// ---- Math natives ----
-
-/// Native: `Math.max(int, int)` — returns the larger value.
-pub(crate) fn native_math_max_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_int_arg(args, 0)?;
-    let b = extract_int_arg(args, 1)?;
-    Ok(Some(Slot::Int(a.max(b))))
-}
-
-/// Native: `Math.min(int, int)` — returns the smaller value.
-pub(crate) fn native_math_min_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_int_arg(args, 0)?;
-    let b = extract_int_arg(args, 1)?;
-    Ok(Some(Slot::Int(a.min(b))))
-}
-
-/// Native: `Math.abs(int)` — returns absolute value.
-pub(crate) fn native_math_abs_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_int_arg(args, 0)?;
-    Ok(Some(Slot::Int(a.wrapping_abs())))
-}
-
-/// Native: `Math.floorMod(int, int)` — remainder with the divisor's sign.
-pub(crate) fn native_math_floor_mod_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_int_arg(args, 0)?;
-    let b = extract_int_arg(args, 1)?;
-    if b == 0 {
-        return Err(Error::DivisionByZero);
-    }
-    let remainder = a.wrapping_rem(b);
-    let floor_mod = if remainder != 0 && (remainder < 0) != (b < 0) {
-        remainder.wrapping_add(b)
-    } else {
-        remainder
-    };
-    Ok(Some(Slot::Int(floor_mod)))
-}
-
-// ---- Extended Math natives ----
-
-/// Native: `Math.sqrt(double)` — returns square root.
-pub(crate) fn native_math_sqrt(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.sqrt())))
-}
-
-/// Native: `Math.pow(double, double)` — returns a raised to the power b.
-pub(crate) fn native_math_pow(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    let b = extract_double_arg(args, 1)?;
-    Ok(Some(Slot::Double(a.powf(b))))
-}
-
-/// Native: `Math.floor(double)` — returns floor value.
-pub(crate) fn native_math_floor(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.floor())))
-}
-
-/// Native: `Math.ceil(double)` — returns ceiling value.
-pub(crate) fn native_math_ceil(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.ceil())))
-}
-
-/// Native: `Math.round(double)` — returns closest long.
-#[allow(clippy::cast_possible_truncation)]
-pub(crate) fn native_math_round_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Long(a.round() as i64)))
-}
-
-/// Native: `Math.abs(long)` — returns absolute value.
-pub(crate) fn native_math_abs_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_long_arg(args, 0)?;
-    Ok(Some(Slot::Long(a.wrapping_abs())))
-}
-
-/// Native: `Math.abs(double)` — returns absolute value.
-pub(crate) fn native_math_abs_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.abs())))
-}
-
-/// Native: `Math.max(long, long)` — returns the larger value.
-pub(crate) fn native_math_max_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_long_arg(args, 0)?;
-    let b = extract_long_arg(args, 1)?;
-    Ok(Some(Slot::Long(a.max(b))))
-}
-
-/// Native: `Math.min(long, long)` — returns the smaller value.
-pub(crate) fn native_math_min_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_long_arg(args, 0)?;
-    let b = extract_long_arg(args, 1)?;
-    Ok(Some(Slot::Long(a.min(b))))
-}
-
-/// Native: `Math.max(double, double)` — returns the larger value.
-pub(crate) fn native_math_max_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    let b = extract_double_arg(args, 1)?;
-    Ok(Some(Slot::Double(a.max(b))))
-}
-
-/// Native: `Math.min(double, double)` — returns the smaller value.
-pub(crate) fn native_math_min_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    let b = extract_double_arg(args, 1)?;
-    Ok(Some(Slot::Double(a.min(b))))
-}
-
-// ---- Extended Math trig / transcendental natives ----
-
-/// Native: `Math.sin(double)` — sine (argument in radians).
-pub(crate) fn native_math_sin(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.sin())))
-}
-
-/// Native: `Math.cos(double)` — cosine (argument in radians).
-pub(crate) fn native_math_cos(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.cos())))
-}
-
-/// Native: `Math.tan(double)` — tangent (argument in radians).
-pub(crate) fn native_math_tan(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.tan())))
-}
-
-/// Native: `Math.asin(double)` — arc sine, result in [-π/2, π/2].
-pub(crate) fn native_math_asin(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.asin())))
-}
-
-/// Native: `Math.acos(double)` — arc cosine, result in [0, π].
-pub(crate) fn native_math_acos(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.acos())))
-}
-
-/// Native: `Math.atan(double)` — arc tangent, result in [-π/2, π/2].
-pub(crate) fn native_math_atan(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.atan())))
-}
-
-/// Native: `Math.atan2(double, double)` — angle of vector (y, x) in [-π, π].
-pub(crate) fn native_math_atan2(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let y = extract_double_arg(args, 0)?;
-    let x = extract_double_arg(args, 1)?;
-    Ok(Some(Slot::Double(y.atan2(x))))
-}
-
-/// Native: `Math.log(double)` — natural logarithm.
-pub(crate) fn native_math_log(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.ln())))
-}
-
-/// Native: `Math.log10(double)` — base-10 logarithm.
-pub(crate) fn native_math_log10(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.log10())))
-}
-
-/// Native: `Math.exp(double)` — Euler's number raised to the given power.
-pub(crate) fn native_math_exp(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.exp())))
-}
-
-/// Native: `Math.signum(double)` — sign of a: -1.0, 0.0, or 1.0.
-pub(crate) fn native_math_signum_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.signum())))
-}
-
-/// Native: `Math.signum(float)` — sign of a as float: -1.0, 0.0, or 1.0.
-#[allow(clippy::cast_possible_truncation)]
-pub(crate) fn native_math_signum_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_float_arg(args, 0)?;
-    Ok(Some(Slot::Float(a.signum())))
-}
-
-/// Native: `Math.toRadians(double)` — converts degrees to radians.
-pub(crate) fn native_math_to_radians(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.to_radians())))
-}
-
-/// Native: `Math.toDegrees(double)` — converts radians to degrees.
-pub(crate) fn native_math_to_degrees(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.to_degrees())))
-}
-
-/// Native: `Math.cbrt(double)` — cube root.
-pub(crate) fn native_math_cbrt(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_double_arg(args, 0)?;
-    Ok(Some(Slot::Double(a.cbrt())))
-}
-
-/// Native: `Math.hypot(double, double)` — sqrt(x²+y²) without overflow.
-pub(crate) fn native_math_hypot(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let x = extract_double_arg(args, 0)?;
-    let y = extract_double_arg(args, 1)?;
-    Ok(Some(Slot::Double(x.hypot(y))))
-}
-
-/// Native: `Math.floorDiv(int, int)` — largest int ≤ quotient.
-pub(crate) fn native_math_floor_div_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_int_arg(args, 0)?;
-    let b = extract_int_arg(args, 1)?;
-    if b == 0 {
-        return Err(Error::DivisionByZero);
-    }
-    Ok(Some(Slot::Int(
-        a.wrapping_div_euclid(b) - i32::from(a.wrapping_rem(b) != 0 && (a < 0) != (b < 0)),
-    )))
-}
-
-/// Native: `Math.round(float)` — rounds float to nearest int.
-#[allow(clippy::cast_possible_truncation)]
-pub(crate) fn native_math_round_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> Result<Option<Slot>> {
-    let a = extract_float_arg(args, 0)?;
-    Ok(Some(Slot::Int(a.round() as i32)))
-}
-
-#[cfg(test)]
-mod havoc_proptest_math {
-    use super::*;
-    use proptest::prelude::*;
-    use std::io::sink;
-
-    proptest! {
-        #[test]
-        fn fuzz_native_math_floor_div_int(a in any::<i32>(), b in any::<i32>()) {
-            let mut heap = duke_gc::Heap::new();
-            let mut control = NativeControl::default();
-            let args = vec![Slot::Int(a), Slot::Int(b)];
-
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                let _ = native_math_floor_div_int(&args, &mut heap, &mut sink(), &mut control);
-            }));
-
-            assert!(result.is_ok(), "Panic on a={a}, b={b}");
-        }
-    }
-}
-
+include!("native/math.rs");
 // ---- System.arraycopy native ----
 
 /// Native: `System.arraycopy(Object src, int srcPos, Object dst, int dstPos, int length)`.
