+++ crates/duke-interpreter/src/native/math.rs
@@ -1,0 +1,418 @@
+// ---- Math natives ----
+
+/// Native: `Math.max(int, int)` — returns the larger value.
+pub(crate) fn native_math_max_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_int_arg(args, 0)?;
+    let b = extract_int_arg(args, 1)?;
+    Ok(Some(Slot::Int(a.max(b))))
+}
+
+/// Native: `Math.min(int, int)` — returns the smaller value.
+pub(crate) fn native_math_min_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_int_arg(args, 0)?;
+    let b = extract_int_arg(args, 1)?;
+    Ok(Some(Slot::Int(a.min(b))))
+}
+
+/// Native: `Math.abs(int)` — returns absolute value.
+pub(crate) fn native_math_abs_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_int_arg(args, 0)?;
+    Ok(Some(Slot::Int(a.wrapping_abs())))
+}
+
+/// Native: `Math.floorMod(int, int)` — remainder with the divisor's sign.
+pub(crate) fn native_math_floor_mod_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_int_arg(args, 0)?;
+    let b = extract_int_arg(args, 1)?;
+    if b == 0 {
+        return Err(Error::DivisionByZero);
+    }
+    let remainder = a.wrapping_rem(b);
+    let floor_mod = if remainder != 0 && (remainder < 0) != (b < 0) {
+        remainder.wrapping_add(b)
+    } else {
+        remainder
+    };
+    Ok(Some(Slot::Int(floor_mod)))
+}
+
+// ---- Extended Math natives ----
+
+/// Native: `Math.sqrt(double)` — returns square root.
+pub(crate) fn native_math_sqrt(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.sqrt())))
+}
+
+/// Native: `Math.pow(double, double)` — returns a raised to the power b.
+pub(crate) fn native_math_pow(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    let b = extract_double_arg(args, 1)?;
+    Ok(Some(Slot::Double(a.powf(b))))
+}
+
+/// Native: `Math.floor(double)` — returns floor value.
+pub(crate) fn native_math_floor(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.floor())))
+}
+
+/// Native: `Math.ceil(double)` — returns ceiling value.
+pub(crate) fn native_math_ceil(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.ceil())))
+}
+
+/// Native: `Math.round(double)` — returns closest long.
+#[allow(clippy::cast_possible_truncation)]
+pub(crate) fn native_math_round_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Long(a.round() as i64)))
+}
+
+/// Native: `Math.abs(long)` — returns absolute value.
+pub(crate) fn native_math_abs_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_long_arg(args, 0)?;
+    Ok(Some(Slot::Long(a.wrapping_abs())))
+}
+
+/// Native: `Math.abs(double)` — returns absolute value.
+pub(crate) fn native_math_abs_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.abs())))
+}
+
+/// Native: `Math.max(long, long)` — returns the larger value.
+pub(crate) fn native_math_max_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_long_arg(args, 0)?;
+    let b = extract_long_arg(args, 1)?;
+    Ok(Some(Slot::Long(a.max(b))))
+}
+
+/// Native: `Math.min(long, long)` — returns the smaller value.
+pub(crate) fn native_math_min_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_long_arg(args, 0)?;
+    let b = extract_long_arg(args, 1)?;
+    Ok(Some(Slot::Long(a.min(b))))
+}
+
+/// Native: `Math.max(double, double)` — returns the larger value.
+pub(crate) fn native_math_max_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    let b = extract_double_arg(args, 1)?;
+    Ok(Some(Slot::Double(a.max(b))))
+}
+
+/// Native: `Math.min(double, double)` — returns the smaller value.
+pub(crate) fn native_math_min_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    let b = extract_double_arg(args, 1)?;
+    Ok(Some(Slot::Double(a.min(b))))
+}
+
+// ---- Extended Math trig / transcendental natives ----
+
+/// Native: `Math.sin(double)` — sine (argument in radians).
+pub(crate) fn native_math_sin(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.sin())))
+}
+
+/// Native: `Math.cos(double)` — cosine (argument in radians).
+pub(crate) fn native_math_cos(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.cos())))
+}
+
+/// Native: `Math.tan(double)` — tangent (argument in radians).
+pub(crate) fn native_math_tan(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.tan())))
+}
+
+/// Native: `Math.asin(double)` — arc sine, result in [-π/2, π/2].
+pub(crate) fn native_math_asin(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.asin())))
+}
+
+/// Native: `Math.acos(double)` — arc cosine, result in [0, π].
+pub(crate) fn native_math_acos(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.acos())))
+}
+
+/// Native: `Math.atan(double)` — arc tangent, result in [-π/2, π/2].
+pub(crate) fn native_math_atan(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.atan())))
+}
+
+/// Native: `Math.atan2(double, double)` — angle of vector (y, x) in [-π, π].
+pub(crate) fn native_math_atan2(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let y = extract_double_arg(args, 0)?;
+    let x = extract_double_arg(args, 1)?;
+    Ok(Some(Slot::Double(y.atan2(x))))
+}
+
+/// Native: `Math.log(double)` — natural logarithm.
+pub(crate) fn native_math_log(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.ln())))
+}
+
+/// Native: `Math.log10(double)` — base-10 logarithm.
+pub(crate) fn native_math_log10(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.log10())))
+}
+
+/// Native: `Math.exp(double)` — Euler's number raised to the given power.
+pub(crate) fn native_math_exp(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.exp())))
+}
+
+/// Native: `Math.signum(double)` — sign of a: -1.0, 0.0, or 1.0.
+pub(crate) fn native_math_signum_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.signum())))
+}
+
+/// Native: `Math.signum(float)` — sign of a as float: -1.0, 0.0, or 1.0.
+#[allow(clippy::cast_possible_truncation)]
+pub(crate) fn native_math_signum_float(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_float_arg(args, 0)?;
+    Ok(Some(Slot::Float(a.signum())))
+}
+
+/// Native: `Math.toRadians(double)` — converts degrees to radians.
+pub(crate) fn native_math_to_radians(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.to_radians())))
+}
+
+/// Native: `Math.toDegrees(double)` — converts radians to degrees.
+pub(crate) fn native_math_to_degrees(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.to_degrees())))
+}
+
+/// Native: `Math.cbrt(double)` — cube root.
+pub(crate) fn native_math_cbrt(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_double_arg(args, 0)?;
+    Ok(Some(Slot::Double(a.cbrt())))
+}
+
+/// Native: `Math.hypot(double, double)` — sqrt(x²+y²) without overflow.
+pub(crate) fn native_math_hypot(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let x = extract_double_arg(args, 0)?;
+    let y = extract_double_arg(args, 1)?;
+    Ok(Some(Slot::Double(x.hypot(y))))
+}
+
+/// Native: `Math.floorDiv(int, int)` — largest int ≤ quotient.
+pub(crate) fn native_math_floor_div_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_int_arg(args, 0)?;
+    let b = extract_int_arg(args, 1)?;
+    if b == 0 {
+        return Err(Error::DivisionByZero);
+    }
+    Ok(Some(Slot::Int(
+        a.wrapping_div_euclid(b) - i32::from(a.wrapping_rem(b) != 0 && (a < 0) != (b < 0)),
+    )))
+}
+
+/// Native: `Math.round(float)` — rounds float to nearest int.
+#[allow(clippy::cast_possible_truncation)]
+pub(crate) fn native_math_round_float(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> Result<Option<Slot>> {
+    let a = extract_float_arg(args, 0)?;
+    Ok(Some(Slot::Int(a.round() as i32)))
+}
+
+#[cfg(test)]
+mod havoc_proptest_math {
+    use super::*;
+    use proptest::prelude::*;
+    use std::io::sink;
+
+    proptest! {
+        #[test]
+        fn fuzz_native_math_floor_div_int(a in any::<i32>(), b in any::<i32>()) {
+            let mut heap = duke_gc::Heap::new();
+            let mut control = NativeControl::default();
+            let args = vec![Slot::Int(a), Slot::Int(b)];
+
+            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
+                let _ = native_math_floor_div_int(&args, &mut heap, &mut sink(), &mut control);
+            }));
+
+            assert!(result.is_ok(), "Panic on a={a}, b={b}");
+        }
+    }
+}
